==> 0.00_KML_converter.Rmd <==
---
title: "KML converter"
author: "Jill Deines"
date: "Sunday, May 22, 2016"
output: 
  html_document:
    toc: yes
---

Goal: Google Earth Engine imports vector data as fusion tables, which must either be .csv or .txt (for points) or .kml (for geometeries).

```{r knitrOpts, echo=FALSE}
library(knitr)
opts_chunk$set(cache=TRUE, cache.path='cache/0.0_kml/',
               fig.path='figure/0.0_kml/',
               eval=FALSE)
```

**R Packages Needed**

```{r packages, message=FALSE, echo=TRUE}
library(rgdal)
library(raster)
```

## Load and convert
Geometries of interest to me are:

* My RRCA + RRB buffered region
* the full HPA with a 30km buffer
* the full HPA with a 30 km buffer + the buffered RRB
* the NE RRB (clipped from full basin)
* the Middle Republican NRD (clipped from RRB NRDs)
* Well Point Datasets (?) - start with RRCA wells for MidRep NRD trials to see if it's helpful
* additional geometries with which to tally irrigated area
* the RRCA model bound

### Load and pre-process

```{r specifyPaths}
# filepaths
gisDir <- 'S:/Users/deinesji/HPA/gis'
#outDir <- 'S:/Users/deinesji/HPA/gis/GEE_processing'
outDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/kml'

# shapefile names of interest (located in gisDir)
hpabuffname <- 'AquiferBoundary_30kmBuffer_WGS84'
rrbBuffname <- 'BigExtent_RRB_RRCA_buff'
rrbstatesname <- 'states_basin_clip'
nrdsname <- 'RRB_NRDs'
hpaRrbBuffname <- 'HPA_30km_RRB_buff'

# load and process if necessary
bigbuff <- readOGR(gisDir, hpaRrbBuffname, verbose=F)
hpabuff <- readOGR(gisDir, hpabuffname, verbose=F)
rrbbuff <- readOGR(gisDir, rrbBuffname, verbose=F)
rrbstates <- readOGR(gisDir, rrbstatesname, verbose=F)
rrbNE <- rrbstates[rrbstates$state == 'NE',]
nrds <- readOGR(gisDir, nrdsname, verbose=F)
nrd1 <- nrds[nrds$NRD_Name == 'Middle Republican',]
rrcaWells <- readOGR(paste0(gisDir,'/Wells'), 'RRCA_NE_wells_2007')

# new bounds to add to summary part (12/5/16)
usstates <- readOGR('S:/Users/deinesji/GIS_general',
                    'States_continental', verbose = F)
usstates <- spTransform(usstates, proj4string(rrbbuff))
rrcaExtentStates <- raster::intersect(rrbbuff, usstates)

# rrca model bound, and by state
rrcaBound <- readOGR(paste0(gisDir,'/RRCA_Model_Export'), 'bound')
rrcaBound$MasterID <- 'RRCA model boundary'

usstates2 <- spTransform(usstates, proj4string(rrcaBound))
rrcaBoundStates <- raster::intersect(rrcaBound, usstates2)
rrcaBoundStates$masterID <- paste0(rrcaBoundStates$STATE_ABBR,'_rrcaBound')

gmdDir <- 'S:/Data/GIS_Data/Downloaded/High_Plains_Aquifer/HPA_Water_Entities/Kansas'
gmds <- readOGR(gmdDir, 'KS - GROUNDWATER_DISTRICTS', verbose=F)
gmd4 <- gmds[gmds$ID == 4,]
#plot(gmd4)
```

### To KML function
Re-projects to WGS84 as required for KML

```{r kmlFun}
toKml <- function(spdf, filename, outDir) {
  # spdf: polygon object in R
  # filename: for output file
  # outDir: output directory
  spdf.wgs84 <- spTransform(spdf, CRS("+proj=longlat +datum=WGS84"))
  writeOGR(spdf.wgs84, paste0(outDir, '/', filename, '.kml'),
            layer = 'layer', driver = 'KML')
}
```

### Write out KMLs

```{r outKML}
toKml(hpabuff, 'HPA_30kmBuff', outDir)
toKml(rrbbuff, rrbBuffname, outDir)
toKml(rrbNE, 'NE_rrb', outDir)
toKml(nrd1, 'NRD_MidRepub', outDir)
toKml(rrcaWells, 'RRCA_NE_Wells_2007', outDir)

toKml(bigbuff, 'HPA_30km_RRB_buff',outDir)

# new stuff for summary trends (note post-hoc addition of masterID column in fusion tables
toKml(rrbstates, 'RRB_basin_states', outDir)
toKml(nrds, 'NE_NRDs_rrb_clip', outDir)
toKml(rrcaExtentStates, 'RRB_fullExtent_states', outDir)
toKml(gmd4, 'KS_GMD4', outDir)
toKml(rrcaBound, 'RRCA_bound', outDir)
toKml(rrcaBoundStates, 'RRCA_bound_states',outDir)

# test <- readOGR(paste0(outDir, '/RRB_basin_states.kml'), 'boundExtent')
```
