==> 1.10_makeValidationPoints.Rmd <==
---
title: "Create Validation Points"
author: "Jill Deines"
date: "Thursday, May 19, 2016"
output: 
  html_document:
    toc: yes
---

Goal: Create a set of validation points to be used in evaluating irrigation maps produced from remote sensing classifications. The aim is to have random points stratified among different crop types as well as non-ag land.

There are three tiers of area interest:

* Middle Republican NRD, NE
* Republican River Basin/RRCA groundwater model
* The rest of the HPA

```{r knitrOpts, echo=FALSE}
library(knitr)
opts_chunk$set(cache=TRUE, cache.path='cache/1.1_vPOints/',
               fig.path='figure/1.1_vPoints/')
```

**R Packages Needed**

```{r packages, message=FALSE, echo=TRUE}
library(rgdal)
library(raster)
library(rgeos)
library(dplyr)

# my functions (to be packaged)
source('functions/getStratCdl.R')
```

## Load Stuff
CDLs downloaded from Cropscape after 05/28/2016 and re-formatted via R to reduce file size. It's AOI is the HPA aquifer + RRCA/RRB buffer (HPA_RRB_buff.zip).

Also load the CDL csv key (the updated, consistent key across all years), and polygons for area extent.

All polygons are re-projected to match CDL projection. Must use a meters based projection for sampling function defined below.


```{r loadStuff}
# ## load original cdl download and write out smaller
#reduce <- raster('F:/deinesji/CDL/zipDownloads/CDL_2007_clip_20161004135148_550013268.tif')
#writeRaster(reduce, 'F:/deinesji/CDL/hpa_rrbbuff/CDL_2007_swNebraska.tif',datatype = 'INT1U')

#cdl and key
#cdl <- raster('S:/Users/deinesji/HPA/gis/CDL/fullHPA/CDL_2010_HPA_RRB.tif')
cdl <- raster('S:/Users/deinesji/HPA/gis/CDL/CDL_RRB_2007_adj.tif')
cdlKey <- read.csv('S:/Users/deinesji/HPA/gis/CDL/CDL_key_2014.csv', 
                   stringsAsFactors=F)

# projection datum not read in for cdl, fix 
tweakProj <- "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0 +datum=NAD83 +units=m +no_defs +ellps=GRS80 +towgs84=0,0,0"
proj4string(cdl) <- tweakProj


# Jill's HPA gis dir
gisDir <- 'S:/Users/deinesji/HPA/gis'


# big RRB/RRCA boundary and transform to AEA proj
rrcabuff <- readOGR(gisDir, 'BigExtent_RRB_RRCA_buff', 
                    verbose=F)
rrcabuff <- spTransform(rrcabuff, CRS(tweakProj))

# Middle Republican RRB (smaller area for testing)
nrds <-  readOGR(gisDir, 'RRB_NRDs', verbose=F)
midRep <- nrds[nrds$NRD_Name == 'Middle Republican',]
midRep <- spTransform(midRep, CRS(tweakProj))

# HPA (no buffer)
hpa <- readOGR(gisDir, 'Aquifer_Regions_dissolve', verbose=F)
hpa <- spTransform(hpa, CRS(tweakProj))
```


## Make Point Datasets

### Middle Republican
I might try evaluating classifications in this NRD

```{r midRep}
# # # crop CDL
#cdl.crop2 <- crop(cdl, midRep)
#cdl.mask2 <- mask(cdl.crop2, midRep)
# # 
# # # write out crop mask so faster next time
#writeRaster(cdl.mask2, dataType = 'INT1U', filename = 'S:/Users/deinesji/HPA/gis/CDL/rsValidationMaskedSets/2007_midRepMask.tif')

# load previously masked
cdl.mask.midRep <- raster('S:/Users/deinesji/HPA/gis/CDL/rsValidationMaskedSets/2007_midRepMask.tif')
proj4string(cdl.mask.midRep) <- tweakProj

# get points!
ptDensity.nrd <- 10 # 1 point per every "x" km2 
grasslandThin <- 0.08 # keep this proportion of grassland points
sampleFactor <- 2.8 # increase sample points to offset grassland and CDL category thinning

# generate stratified random sample
nrdPoints <- getStratCdl(cdl.mask.midRep, midRep, ptDensity.nrd, sampleFactor,
                         grasslandThin, cdlKey, seed=1)

# see how points cover classes
table(nrdPoints@data$CLASS_NAME)
  
dim(nrdPoints)
  
# view
plot(cdl.mask.midRep)
plot(nrdPoints, add=T, pch = 19)

# see how close to target density after grassland/developed thinning
dim(nrdPoints)[1]
gArea(midRep)*1e-6/ptDensity.nrd
```

Note: alfalfa is low to reliably get samples for irrigated/non-irrigated alfalfa

```{r nrdWrite, eval=FALSE}
# write out as WGS84 (needed for kml)
nrdPoints.ll <- spTransform(nrdPoints, CRS("+proj=longlat +datum=WGS84"))
outdir <- 'S:/Users/deinesji/HPA/gis/GEE_processing/validPoints'
writeOGR(nrdPoints.ll, outdir, 'valid_MidRep_10_2007_20161011', driver='ESRI Shapefile', 
         overwrite = F)
writeOGR(nrdPoints.ll, paste0(outdir, '/valid_MidRep_10_2007_20161011.kml'),
            layer = 'thislayer', driver='KML', overwrite=F)

# write KML to my dropbox for laptop access 
writeOGR(nrdPoints.ll, 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/validation/generatedPointsForUpload/valid_MidRep_10_2007_20161011.kml', layer = 'thislayer', driver='KML', overwrite=F)
```

### Republican River Basin
including RRCA model domain

```{r rrca, eval=FALSE}
# crop CDL
cdl.crop3 <- crop(cdl, rrcabuff)
cdl.mask3 <- mask(cdl.crop3, rrcabuff)

# get points!
ptDensity <- 100 # 1 point per every 100 km2 

# time run
start.time <- Sys.time()

rrcaPoints <- getStratCdl(cdl.mask3, rrcabuff, ptDensity, cdlKey, seed=1)
  
end.time <- Sys.time()
time.taken <- end.time - start.time
```


### Full Study Area
Extent of the High Plains Aquifer + the full Republican River Basin (some sticks out)

**Prior to generating points for the full area, I will likely do additional masks of urban areas, etc, to increase point density on areas of interest**

```{r fullPoints, eval=FALSE}
# union the HPA extent and RRCA buffer
fullArea <- gUnion(hpa, rrcabuff)

# crop CDL
cdl.crop <- crop(cdl, fullArea)
cdl.mask <- mask(cdl.crop, fullArea)

# get points!
ptDensity <- 200 # 1 point per every 100 km2 

# time run
start.time <- Sys.time()

fullPoints <- getStratCdl(cdl.mask, fullArea, ptDensity, cdlKey, seed=1)
  
end.time <- Sys.time()
time.taken <- end.time - start.time
```


## Old Development Process and Brainstorming
I want stratified random samples based on crop classes and non-crop. I could make a point for every raster cell, convert that to a data frame, and then take stratified samples from there. This would also allow me to make a few different sized sets of points.

But then points are regular, not random. Make random points, then stratify?

## Middle Republican NRD test

### Method 1: Random points
Generate random points, extract raster CDL values, tabulate

(No control over stratification - could generate more points and thin)

I use method 2.

```{r midRep_test, eval=FALSE}
# Middle Republican RRB (smaller area for testing)
nrds <-  readOGR('S:/Users/deinesji/HPA/gis', 'RRB_NRDs', verbose=F)
midRep <- nrds[nrds$NRD_Name == 'Middle Republican',]
midRep <- spTransform(midRep, CRS(tweakProj))

# get area of buffer
nrdArea <- gArea(midRep)
nrdkm2 <- nrdArea * 1e-6
nrdkm2  # square kilometers

# desired point numbers to mimic density in Brown and Pervez 2010
nrdkm2/100 # min number of points?
nrdkm2/20 # max number of points

# generate random points
npoints <- 500
p1 <- spsample(midRep, n =npoints, "random")
plot(midRep)
plot(p1, add=T, pch=19)

# conver to spdf
p1 <- SpatialPointsDataFrame(p1, data=data.frame(ID=1:npoints))

# add CDL crop type from raster and category name from key
p1$CDLcode <- extract(cdl, p1)
p1 <- merge(p1, cdlKey[,c('VALUE','CLASS_NAME')], 
            by.x='CDLcode', by.y='VALUE', all.y = F)

# tabulate to see class representations
classTab <- table(p1$CLASS_NAME)
```


### Method 2: Raster to Points (use this)
Convert Raster to a point dataset -> data.frame, and do stratified random sampling from there

```{r methodTest2, eval=FALSE}
# crop CDL to NRD
cdl.mr <- crop(cdl, midRep)

# convert to df
r.points <- as.data.frame(cdl.mr, xy = T, centroids=T)
names(r.points) <- c('x','y','CDLcode')

# take a proportional sample from all classes
set.seed(1)
p1.r <- r.points %>%
          group_by(CDLcode) %>%
          sample_frac(size = 4/100000)

# add crop key (do after sampling if final method)
r.points <- merge(p1.r, cdlKey[,c('VALUE','CLASS_NAME')], 
            by.x='CDLcode', by.y='VALUE', all.y = F)

classTab.r <- table(p1.r$CLASS_NAME)
```
