==> 1.16_trainingPoints_RRB.Rmd <==
---
title: "Training Points, rRB"
author: "Jill Deines"
date: "Monday, November 28, 2016"
output: 
  html_document:
    toc: yes
---

Goal: Merge training points exports from GEE for each year (starting with 2010 and 2012) and make cool plots of greenness time series.

```{r knitrOpts, echo=FALSE}
library(knitr)
opts_chunk$set(cache=FALSE, cache.path='cache/1.16_trainRRB/',
               fig.path='figure/1.16_trainRRB/')
```

**R Packages Needed**

```{r packages, message=FALSE, echo=TRUE}
library(rgdal)
library(latticeExtra)
library(ggplot2)
library(tidyr)
```

## Load and Merge Data
training points with anthony vars included = v02

```{r loadData, fig.width = 5.5, fig.height=4.5, cache=F, dpi=600, dev=c('png','pdf')}
# file directory and filenames
trainFolder <- 'C:/Users/deinesji/Google Drive/GEE_validation/trainingSets'
files <- list.files(trainFolder, full.names = T, pattern='*_v02.csv')

# load and merge
trainList <- list()
for (i in 1:length(files)){
  trainList[[i]] <- read.csv(files[i]) 
}
trainAll <- do.call("rbind",trainList)

# number of total training points
nrow(trainAll)

# spatialize
trainPoints <- trainAll
coordinates(trainPoints) <- ~ long + lat
proj4string(trainPoints) <- CRS("+proj=longlat +datum=WGS84")

# Plot with Koepen counties
tempDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/validation/1.15_makeRrbPoints_workspace'
koep <- readOGR(tempDir, 'Koep_RRB_dissolve', verbose=F)

# basin boundaries
gisDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/kml'
rrb <- readOGR(paste0(gisDir,'/BigExtent_RRB_RRCA_buff.kml'), 'boundExtent', verbose=F)


#par(mar=c(0,0,0,0))
koepPal <- c('#eff3ff','#bdd7e7','#6baed6')
pch2 <- 1 # 19 solid, 1 open, 3 plus, 4 x)
ptcex <- 1
plot(koep, col=koepPal, axes=T)
plot(trainPoints[trainPoints$year == 2010,], col='#fd8d3c',add=T, cex=ptcex,
     pch=pch2, axes=F)
plot(trainPoints[trainPoints$year == 2012,], col='#cb181d',add=T, cex=ptcex,
     pch=pch2, axes=F)
legend('topright', legend=as.character(koep$koep), fill=koepPal, bty='n',
       cex=1)
legend('topleft', legend=c('2010','2012'), col=c('#fd8d3c','#cb181d'),
       bty='n', pch=pch2,cex=1)


```

## Export training sets for GEE
This exports combined points as KML's for use in classification in GEE. Exports include:

* 2010 combined points for the RRB
* 2012 combined points for the RRB
* 2010 + 2012 combined points for the RRB

First, I will try exporting points with the full set of variables in tact, in hopes that it will eliminate re-sampling imagery in the classification step. Therefore, I remove all variables except the numerical categories (masterType) and the image values.

```{r exportForGee, eval = FALSE}
exportFolder <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/training'

# use a crop/irrigated status crossed class type to maximize downstream flexibility
trainPoints$masterType <- paste0(trainPoints$crop,'_',trainPoints$class)
masterKey <- data.frame(masterType = sort(unique(trainPoints$masterType)),
                        masterNum = 0:(length(unique(trainPoints$masterType))-1))
trainPoints1 <- merge(trainPoints, masterKey)

# clean up data
trainExport <- trainPoints1[ , -which(names(trainPoints1) %in% 
                c(".geo",'certainty', 'system.index','class','crop','koep','classNum'))]

# split up points
train2010 <- trainExport[trainExport$year == 2010,]
train2012 <- trainExport[trainExport$year == 2012,]

# remove year column
train2010 <- train2010[ , -which(names(train2010) == 'year')]
train2012 <- train2012[ , -which(names(train2012) == 'year')]                       
trainExport1 <- trainExport[ , -which(names(trainExport) == 'year')]                      

# write out KML's
writeOGR(train2010, paste0(exportFolder, '/2010_training_12class_v2.kml'),
         layer = 'layer', driver = 'KML')
writeOGR(train2012, paste0(exportFolder, '/2012_training_12class_v2.kml'),
         layer = 'layer', driver = 'KML')
writeOGR(trainExport1, paste0(exportFolder, '/2010_2012_COMBINED_training_12class_v2.kml'),
         layer = 'layer', driver = 'KML')

# write out corresponding CSV's for random forest work in R
write.csv(trainExport1@data, row.names = F,
          paste0(exportFolder,
                 '/2010_2012_COMBINED_training_12class_v2.csv'))

```


## Visualize

### format dataframe for visualization

```{r visFormatting}
# remove .geo column
trainAll <- trainAll[ , -which(names(trainAll) %in% 
                      c(".geo",'certainty', 'system.index', 'lat','long'))]

# change from wide to long format
trainAllLong <- gather(trainAll, key = rasValue, value = value, 
                       EVI_max_10:slope_mean, # combine these columns
                       # except for these columns
                       -c(class,crop,koep,classNum))

# add a month/year timestamp
monthvector <- as.numeric(unlist(regmatches(trainAllLong$rasValue, 
                          gregexpr("[[:digit:]]+", trainAllLong$rasValue))))
monthvectorPlus <- c(monthvector, rep(0, nrow(trainAllLong) - length(monthvector)))
trainAllLong$month <- monthvectorPlus

# add a master category
trainAllLong$masterType <- paste0(trainAllLong$crop,'_',trainAllLong$class)

# point count by crop type (per year)
train2010 <- trainAllLong[grepl('GI',trainAllLong$rasValue) 
                       & trainAllLong$month != 14
                       & trainAllLong$year == 2010,]
# 2010 breakdown
table(train2010$masterType)/8

train2012 <- trainAllLong[grepl('GI',trainAllLong$rasValue) 
                       & trainAllLong$month != 14
                       & trainAllLong$year == 2012,]
# 2012 breakdown
table(train2012$masterType)/8
```

```{r plotHelperFunctions}
# function to unnest column list from aggregate
flatter<-function (x, recursive = TRUE) {
   dfcolumns <- vapply(x, is.matrix, logical(1))
   cbind(x[!dfcolumns], do.call(cbind, x[dfcolumns]))
}

# function to get mean/stdev for each crop type/irrigation status
# for each month/year for any index or year
aggregateIndexByCrop <- function(index, year, df){
  # select by index and year
  df2 <- df[grepl(index,df$rasValue) & df$year == year
                       & df$month != 14
                       & df$crop != 'fallow',]
  df_agg <- stats::aggregate(value ~ crop + class + month + year, data = df2, 
                               FUN = function(x) {c(mean = mean(x), sd = sd(x))})
  # apply function
  dfPlot <- flatter(df_agg)
}  


# for quartile ribbons
aggregateIndexByCrop_quartiles <- function(index, year, df){
  # select by index and year
  df2 <- df[grepl(index,df$rasValue) & df$year == year
                       & df$month != 14
                       & df$crop != 'fallow',]
  df_agg <- stats::aggregate(value ~ crop + class + month + year, data = df2, 
                               FUN = 'summary')
  # apply function
  dfPlot <- flatter(df_agg)
  
  names(dfPlot)[6] <- 'Q1'
  names(dfPlot)[9] <- 'Q3'
    
  return(dfPlot)
}  


# function to plot irrigation status by crop type facet
plotIndexByCrop <- function(dfPlot, title, index){
  ggplot(data = dfPlot,
         aes(x = month, y = mean, colour = class, group = class)) +
       geom_ribbon(aes(ymin = mean - sd, 
                       ymax = mean + sd, 
                       fill = class), alpha = 0.2, colour=NA) +
      scale_color_manual(values=c('gold2','slateblue2','saddlebrown')) +
      scale_fill_manual(values=c('gold2','slateblue2','saddlebrown')) + 
       geom_line() +
       facet_wrap(~crop) +
       ylab(paste0(index, ' Index')) + 
       xlab('Month (April - November)') + theme_bw() +
       ggtitle(title)
}


# quartiles: function to plot irrigation status by crop type facet
plotIndexByCrop_quartiles <- function(dfPlot, title, index){
  ggplot(data = dfPlot,
         aes(x = month, y = Mean, colour = class, group = class)) +
       geom_ribbon(aes(ymin = Q1, 
                       ymax = Q3, 
                       fill = class), alpha = 0.2, colour=NA) +
      scale_color_manual(values=c('gray50','darkgreen','saddlebrown'),
                         labels=c('Rainfed','Irrigated','Non-crop')) +
      scale_fill_manual(values=c('gray50','darkgreen','saddlebrown'),
                        labels=c('Rainfed','Irrigated','Non-crop')) +
        geom_line() +
        geom_line(aes(x=month, y = Q1, colour = class, group = class),
                  size=.1, alpha=.8) +
         geom_line(aes(x=month, y = Q3, colour = class, group = class),
                  size=.1, alpha=.8) +
       facet_wrap(~crop) +
       ylab(paste0(index, ' Index')) + 
       xlab('Month (April - November)') + theme_bw() +
       theme(text = element_text(size = 16)) 
}


# 1 index per year plot functions for combo vars

subsetIndexByCrop <- function(index, year, df){
  # select by index and year
  df2 <- df[grepl(index,df$rasValue) & df$year == year & df$crop != 'fallow',]
}  


plotCombo <- function(dfPlot, title, index){
  ggplot(data = dfPlot,
         aes(x = crop, y = value, colour = class)) +
       geom_boxplot(aes(colour = class), position = 'dodge') +
       ylab(paste0(index, ' Index')) + 
       xlab('crop') + theme_bw() +
       ggtitle(title)
}

```

### Plot: all messy
A plot of irrigated status by crop facet, both years

```{r plotAllMessy}
index <- 'GI'

# select by index and remove fallow
gi <- trainAllLong[grepl(index,trainAllLong$rasValue) 
                       & trainAllLong$month != 14
                       & trainAllLong$crop != 'fallow',]
table(gi$masterType)/8

# aggregate mean and st.dev by crop/status, month, and year
gi_agg <- stats::aggregate(value ~ crop + class + month + year, data = gi, 
                               FUN = function(x) {c(mean = mean(x), sd = sd(x))})
# unnest list
giPlot <- flatter(gi_agg)

# plot all together
ggplot(data = giPlot,
       aes(x = month, y = mean, 
           colour = interaction(year,class),
           group = interaction(year, class))) +
       geom_ribbon(aes(ymin = mean - sd, 
                       ymax = mean + sd, 
                       fill = interaction(year,class)), alpha = 0.2) +
       geom_line() +
       facet_wrap(~crop) +
       ylab('Greenness Index') + xlab('month') + theme_bw() +
       ggtitle("2010 + 2012 Greenness Index (Green/NIR), full RRB")

# aggregate across years! ----------------------------------------------------
# aggregate mean and st.dev by crop/status, month, and year
gi_agg <- stats::aggregate(value ~ crop + class + month, data = gi, 
                               FUN = function(x) {c(mean = mean(x), sd = sd(x))})
# unnest list
giPlot <- flatter(gi_agg)

# plot all together
ggplot(data = giPlot,
       aes(x = month, y = mean, 
           colour = class,
           group = class)) +
       geom_ribbon(aes(ymin = mean - sd, 
                       ymax = mean + sd, 
                       fill = class), alpha = 0.2) +
       geom_line() +
       facet_wrap(~crop) +
       ylab('Greenness Index') + xlab('month') + theme_bw() +
       ggtitle("2010 + 2012 Combined GI (Green/NIR), full RRB")
```

### Greenness Index Plots by Year - poster fig
updated March 17 2017 for INFEWS meeting

```{r greenIndexPlotByCrop_poster, fig.width = 6.75, fig.height = 4.75, dev='png'}
# Greenness Index, 2010: Quartiles for talk 
dfagg <- aggregateIndexByCrop_quartiles(index = 'GI', year = 2010, df=trainAllLong)
plotIndexByCrop_quartiles(dfPlot = dfagg, title = '2010 Greenness Index (GI)', index = 'Greenness')
```

```{r greenIndexPlotByCrop_manuscript_top_noGrid, fig.width = 6.5, fig.height = 4, dpi = 600, dev=c('png','pdf')}
# Greenness Index, 2010: Quartiles for talk 
dfagg <- aggregateIndexByCrop_quartiles(index = 'GI', year = 2010, df=trainAllLong)
#plotIndexByCrop_quartiles(dfPlot = dfagg, title = '2010 Greenness Index (GI)', index = 'Greenness')

# quartiles: function to plot irrigation status by crop type facet
dfPlot <- dfagg
index <- 'Green'

# adjust facet labels
# plotty plot
labels <- c(`alfalfa` = 'Alfalfa',
            `corn` = 'Corn',
            `noncrop` = 'Non-crop',
            `sorghum` = 'Sorghum',
            `soy` = 'Soy',
            `wheat` = 'Wheat')

ggplot(data = dfPlot,
         aes(x = month, y = Mean, colour = class, group = class)) +
       geom_ribbon(aes(ymin = Q1, 
                       ymax = Q3, 
                       fill = class), alpha = 0.2, colour=NA) +
      scale_color_manual(values=c('gray50','darkgreen','saddlebrown'),
                         labels=c('Rainfed','Irrigated','Non-crop')) +
      scale_fill_manual(values=c('gray50','darkgreen','saddlebrown'),
                        labels=c('Rainfed','Irrigated','Non-crop')) +
        geom_line() +
        geom_line(aes(x=month, y = Q1, colour = class, group = class),
                  size=.1, alpha=.8) +
         geom_line(aes(x=month, y = Q3, colour = class, group = class),
                  size=.1, alpha=.8) +
       facet_wrap(~crop, labeller = as_labeller(labels)) +
       ylab(paste0(index, ' Index')) + 
       xlab('Month (April - November)') + theme_bw() +
       theme(text = element_text(size = 12), 
             legend.position = 'top',
             legend.title = element_blank(),
             panel.grid.major = element_blank(),
             panel.grid.minor = element_blank()) 

```

### Index Plots by Year - talk fig
```{r indexPlotsByCrop, fig.width = 6.75, fig.height = 4.75}
# Greenness Index, 2010: Quartiles for talk 
dfagg <- aggregateIndexByCrop_quartiles(index = 'GI', year = 2010, df=trainAllLong)
plotIndexByCrop_quartiles(dfPlot = dfagg, title = '2010 Greenness Index (GI)', index = 'GI')

# Greenness Index, 2010  
dfagg <- aggregateIndexByCrop(index = 'GI', year = 2010, df=trainAllLong)
plotIndexByCrop(dfPlot = dfagg, title = '2010 GI', index = 'GI')

# Greenness Index, 2012  
dfagg <- aggregateIndexByCrop(index = 'GI', year = 2012, df=trainAllLong)
plotIndexByCrop(dfPlot = dfagg, title = '2012 GI', index = 'GI')

# EVI, 2010  
dfagg <- aggregateIndexByCrop(index = 'EVI', year = 2010, df=trainAllLong)
plotIndexByCrop(dfPlot = dfagg, title = '2010 EVI', index = 'EVI')

# EVI, 2012  
dfagg <- aggregateIndexByCrop(index = 'EVI', year = 2012, df=trainAllLong)
plotIndexByCrop(dfPlot = dfagg, title = '2012 EVI', index = 'EVI')

# NDVI, 2010  
dfagg <- aggregateIndexByCrop(index = 'NDVI', year = 2010, df=trainAllLong)
plotIndexByCrop(dfPlot = dfagg, title = '2010 NDVI', index = 'NDVI')

# NDVI, 2012  
dfagg <- aggregateIndexByCrop(index = 'NDVI', year = 2012, df=trainAllLong)
plotIndexByCrop(dfPlot = dfagg, title = '2012 NDVI', index = 'NDVI')

# NDWI, 2010  
dfagg <- aggregateIndexByCrop(index = 'NDWI', year = 2010, df=trainAllLong)
plotIndexByCrop(dfPlot = dfagg, title = '2010 NDWI', index = 'NDWI')

# NDWI, 2012  
dfagg <- aggregateIndexByCrop(index = 'NDWI', year = 2012, df=trainAllLong)
plotIndexByCrop(dfPlot = dfagg, title = '2012 NDWI', index = 'NDWI')
```


```{r comboVars}
# greenArid 2010
dfsub <- subsetIndexByCrop(index = 'greenArid', year = 2010, df=trainAllLong)
plotCombo(dfPlot = dfsub, title = '2010 Greenness Aridity', index = 'greenness/Arid')

# greenArid 2012
dfsub <- subsetIndexByCrop(index = 'greenArid', year = 2012, df=trainAllLong)
plotCombo(dfPlot = dfsub, title = '2012 Greenness Aridity', index = 'greenness/Arid')

# ndwi*gi 2010
dfsub <- subsetIndexByCrop(index = 'ndwi_gi', year = 2010, df=trainAllLong)
plotCombo(dfPlot = dfsub, title = '2010 NDWI * GI', index = 'ndwi*gi')

# ndwi*gi 2012
dfsub <- subsetIndexByCrop(index = 'ndwi_gi', year = 2012, df=trainAllLong)
plotCombo(dfPlot = dfsub, title = '2012 NDWI * GI', index = 'ndwi*gi')

```

### Plots By Irrigation Status
Combine crop types to compare irrigated, dryland, and noncrop curves
```{r plotByStatus}
index <- 'GI'

# select by index and remove fallow
gi <- trainAllLong[grepl(index,trainAllLong$rasValue) 
                       & trainAllLong$month != 14
                       & trainAllLong$crop != 'fallow',]

# aggregate mean and st.dev by crop/status, month, and year
gi_agg <- stats::aggregate(value ~ class + month + year, data = gi, 
                               FUN = function(x) {c(mean = mean(x), sd = sd(x))})
# unnest list
giPlot <- flatter(gi_agg)

# plot all together
ggplot(data = giPlot,
       aes(x = month, y = mean, 
           colour = class,
           group = class)) +
       geom_ribbon(aes(ymin = mean - sd, 
                       ymax = mean + sd, 
                       fill = class), alpha = 0.2) +
       geom_line() +
       facet_wrap(~year) +
       ylab('Greenness Index') + xlab('month') + theme_bw() +
       ggtitle("2010 + 2012 Greenness Index (Green/NIR), full RRB")
```

### Plots By Koeppengeiger

```{r plotByKoep}
index <- 'GI'

# select by index and remove fallow
gi <- trainAllLong[grepl(index,trainAllLong$rasValue) 
                       & trainAllLong$month != 14
                       & trainAllLong$crop != 'fallow',]

# aggregate mean and st.dev by crop/status, month, and year
gi_agg <- stats::aggregate(value ~ class + month + year + koep, data = gi, 
                               FUN = function(x) {c(mean = mean(x), sd = sd(x))})
# unnest list
giPlot <- flatter(gi_agg)

# plot all together
ggplot(data = giPlot,
       aes(x = month, y = mean, 
           colour = class,
           group = class)) +
       geom_ribbon(aes(ymin = mean - sd, 
                       ymax = mean + sd, 
                       fill = class), alpha = 0.2) +
       geom_line() +
       facet_wrap(~year + koep) +
       ylab('Greenness Index') + xlab('month') + theme_bw() +
       ggtitle("2010 + 2012 Greenness Index (Green/NIR), full RRB")
```
