==> 1.20_validationPointsPostProcess.Rmd <==
---
title: "Validation Points - Post-processing"
author: "Jill Deines"
date: "Sunday, May 29, 2016"
output: 
  html_document:
    toc: yes
---

Goal: Take the KML exported from GEE during validation point creation and post-process. yep.

**Update Sept. 7 2016**: This now has several iterations as the validation point methodology evolved. 

to do: re-write to use helper csvToSpdf function, and make sure to remove system.index


```{r knitrOpts, echo=FALSE}
library(knitr)
opts_chunk$set(cache=TRUE, cache.path='cache/1.2_vPOintsPost/',
               fig.path='figure/1.2_vPointsPost/')
```

**R Packages Needed**

```{r packages, message=FALSE, echo=TRUE}
library(rgdal)
library(rgeos)
```

## Functions

### Process validation points: function
Wrote out imported GEE points as a csv (since the kml write=out doesn't include a layer name). This extracts coordinates, adds crop class name, and turns back into points to write out as a KML to add to GEE as a feature collection.

The following updated function version works for valid3 and valid4 methodology (class numbers and certainty levels).

```{r csvFun}
# testFile <- 'C:/Users/deinesji/Google Drive/GEE_validation/valid_midRep_10_2010_JMDtest.csv'
# cdlKey <- read.csv('S:/Users/deinesji/HPA/gis/CDL/CDL_key_2014.csv', 
#                    stringsAsFactors=F)

csvToSpdf <- function(csvFilename, cdlKey){
    # load output from GEE
    validated <- read.csv(csvFilename, stringsAsFactors=F)

    # extract coordinates from messy .geo string
    coords <- sub(".*\\[(.*)\\].*", "\\1", validated$.geo, perl=TRUE)
    coords2 <- do.call(rbind, strsplit(coords, ','))   
    validated2 <- data.frame(class = validated$class,
                             classNum = validated$classNum,
                             certainty = validated$certainty,
                             CDLcode = validated$max,
                             x = as.numeric(coords2[,1]),
                             y = as.numeric(coords2[,2]),
                             system.index = validated$system.index)

    # add CDL class name
    validated3 <- merge(validated2, cdlKey[,c('VALUE','CLASS_NAME')], 
                    by.x = 'CDLcode', by.y='VALUE', all.y=F)
   
    # spatialize points
    coordinates(validated3) <- ~ x + y
    proj4string(validated3) <- "+proj=longlat +datum=WGS84"
    
    return(validated3)
}

#test <- csvToSpdf(testFile, cdlKey)

```

Keep a version of the original function to be able to run with previous validation sets. You know, for posterity.

```{r oldCsvFun}
# testFile <- 'C:/Users/deinesji/Google Drive/GEE_validation/valid_midRep_10_2010_JMDtest.csv'
# cdlKey <- read.csv('S:/Users/deinesji/HPA/gis/CDL/CDL_key_2014.csv', 
#                    stringsAsFactors=F)

csvToSpdf_old <- function(csvFilename, cdlKey){
    # load output from GEE
    validated <- read.csv(csvFilename, stringsAsFactors=F)

    # extract coordinates from messy .geo string
    coords <- sub(".*\\[(.*)\\].*", "\\1", validated$.geo, perl=TRUE)
    coords2 <- do.call(rbind, strsplit(coords, ','))   
    validated2 <- data.frame(irrigated = validated$irrigated,
                             CDLcode = validated$max,
                             x = as.numeric(coords2[,1]),
                             y = as.numeric(coords2[,2]),
                             system.index = validated$system.index)

    # add CDL class name
    validated3 <- merge(validated2, cdlKey[,c('VALUE','CLASS_NAME')], 
                    by.x = 'CDLcode', by.y='VALUE', all.y=F)
   
    # spatialize points
    coordinates(validated3) <- ~ x + y
    proj4string(validated3) <- "+proj=longlat +datum=WGS84"
    
    return(validated3)
  }

#test <- csvToSpdf(testFile, cdlKey)

```


## Csv to KMl
This section converts datasets exported from GEE as .csv's to spdfs using above function, then writes out as KML's for future import/use as fusion tables in GEE.

Note that validation files prior to the 'valid3' versions use the old csvToSpdf function version

```{r csvToKML_valid1, eval=FALSE}
cdlKey <- read.csv('C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/helper/CDL_key_2014.csv', stringsAsFactors=F)

# Jeremy's first full validation set for NE MidRep NRD
# irrigated and nonirrigated points only (uncertain omitted from export)
fileDir <- 'C:/Users/deinesji/Google Drive/GEE_validation/'
valid1 <- paste0(fileDir, 'valid_midRep_10_2010_jrapp_noUncertainty.csv')
valid1.spdf <- csvToSpdf_old(valid1, cdlKey)
# remove built/water classes
remove <- c(63, 82, 83, 111, 121, 122, 123, 124, 141, 142, 143)
valid1.spdf <- valid1.spdf[!(valid1.spdf$CDLcode %in% remove),]
# make class2 column
noncropCdl <- c(64, 87, 88, 152, 176, 190, 195)
valid1.spdf$class2 <- 1
for (i in 1:nrow(valid1.spdf)){
          if(valid1.spdf$irrigated[i] == "N") valid1.spdf[i,'class2'] <- 0
          if(valid1.spdf$CDLcode[i] %in% noncropCdl) valid1.spdf[i,'class2'] <- 2
  }
# remove system.index column
valid1.spdf<- valid1.spdf[,-3]

writeOGR(valid1.spdf, paste0(fileDir, 'valid_midRep_10_2010_jrapp_noUncertainty.kml'),
            layer = 'validPoints', driver = 'KML')
```


## Point Validation Evaluation
Undergrad Jeremy Rapp evaluated ~980 random points in the Middle Republican NRD. To compare irrigated/nonirrigated classifications between investigators, Jill recorded 222 points. This compares them.

### First trial: binary classification approach


```{r compare_v1}
# crop key
cdlKey <- read.csv('C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/helper/CDL_key_2014.csv', stringsAsFactors=F)

# Jeremey's file
rappFile <-'C:/Users/deinesji/Google Drive/GEE_validation/valid_midRep_10_2010_jrapp.csv'
# Jill's file
jillFile <- 'C:/Users/deinesji/Google Drive/GEE_validation/valid_midRep_10_2010_JMDtest.csv'

# load/process csvs to spatial points
rappPoints <- csvToSpdf_old(rappFile, cdlKey)
jillPoints <- csvToSpdf_old(jillFile, cdlKey)

# project to meters
rappPoints <- spTransform(rappPoints, CRS('+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=37.5 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs '))
jillPoints <- spTransform(jillPoints, CRS('+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=37.5 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs '))

# find distances between all points
distPairs <- as.data.frame(gDistance(jillPoints, rappPoints, byid = T))
distPairs$cellnum <- 1:nrow(distPairs)

# best guess at paired points
pointBuddies <- data.frame(jill = 1:222, rapp = NA, crop = jillPoints$CLASS_NAME,
                           jillIrr = jillPoints$irrigated, rappIrr = NA)
for (i in 1:222) {
  pointBuddies[i,'rapp'] <- distPairs[which.min(distPairs[,i]),'cellnum']
}
  
# add irrigation status for rapp
rappPoints$ind <- 1:nrow(rappPoints)
pointBuds <- merge(pointBuddies, rappPoints[,c('irrigated','ind')],
                   by.x = 'rapp', by.y='ind', all.y = F)
pointBuds$rappIrr <- pointBuds$irrigated

# summarize agreement
totalMatching <- sum(pointBuds$rappIrr == pointBuds$jillIrr)
# percent matching
totalMatching/222

# tabulate crop
pointBuds$identical <- pointBuds$rappIrr == pointBuds$jillIrr
table(pointBuds[,c('crop','identical')])
```


### Second trial: 3-level uncertainty approach
First evaluation on August 22, 2016: neither Jeremy or Jill finished whole set


```{r compare_v3}
# crop key
cdlKey <- read.csv('C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/helper/CDL_key_2014.csv', stringsAsFactors=F)

# Jeremey's file
rappFile <-'C:/Users/deinesji/Google Drive/GEE_validation/valid3_midRep_10_2010_jrappTest.csv'
# Jill's file
jillFile <- 'C:/Users/deinesji/Google Drive/GEE_validation/valid3_midRep_10_2010_jmdTest.csv'

# load/process csvs to spatial points
rappPoints <- csvToSpdf(rappFile, cdlKey)
jillPoints <- csvToSpdf(jillFile, cdlKey)

# project to meters
rappPoints <- spTransform(rappPoints, CRS('+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=37.5 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs '))
jillPoints <- spTransform(jillPoints, CRS('+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=37.5 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs '))

# find distances between all points
distPairs <- as.data.frame(gDistance(jillPoints, rappPoints, byid = T))
distPairs$cellnum <- 1:nrow(distPairs)

# best guess at paired points
pointBuddies <- data.frame(jill = 1:nrow(jillPoints), rapp = NA, distance = NA,
                           crop = jillPoints$CLASS_NAME,
                           jillclass = jillPoints$class, 
                           rappclass = NA,
                           jillcert = jillPoints$certainty,
                           rappcert = NA)

for (i in 1:nrow(jillPoints)) {
  pointBuddies[i,'rapp'] <- distPairs[which.min(distPairs[,i]),'cellnum']
  pointBuddies[i, 'distance'] <- distPairs[which.min(distPairs[,i]),i]
}
  
# add irrigation status (class) and certainty for rapp
rappPoints$ind <- 1:nrow(rappPoints)
pointBuds <- merge(pointBuddies, rappPoints[,c('class','certainty','ind')],
                   by.x = 'rapp', by.y='ind', all.y = F)
pointBuds$rappclass <- pointBuds$class
pointBuds$rappcert <- pointBuds$certainty

# filter for reasonable distances
pointBuds <- pointBuds[pointBuds$distance < 200,]

# summarize agreement
totalMatching <- sum(pointBuds$rappclass == pointBuds$jillclass)
# percent matching
totalMatching/nrow(pointBuds)

# tabulate irrigated/non-irrigated by crop
pointBuds$identical <- pointBuds$rappclass == pointBuds$jillclass
table(pointBuds[,c('crop','identical')])

# check just for certainty level 3
certainPoints <- pointBuds[pointBuds$jillcert == 3 & pointBuds$rappcert ==3,]
matching <- sum(certainPoints$rappclass == certainPoints$jillclass)
matching/nrow(certainPoints)

certainPoints$identical <- certainPoints$rappclass == certainPoints$jillclass
table(certainPoints[,c('crop','identical')])

# write out shapefiles for further investigations
outdir <- 'C:/Users/deinesji/Google Drive/GEE_validation/testShapefiles/secondTest_uncertainty'
#writeOGR(jillPoints, outdir, 'jillpoints_20160822', driver='ESRI Shapefile')
#writeOGR(rappPoints, outdir, 'rapppoints_20160822', driver='ESRI Shapefile')
```


### Third trial: 2-level uncertainty
First evaluation on Sept 7, 2016: Jill using the 2-level uncertainty methods in valid4, compared with Jill's previous points using hte similar 3-level uncertainty methods in valid3


```{r compare_v4}
# crop key
cdlKey <- read.csv('C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/helper/CDL_key_2014.csv', stringsAsFactors=F)

# Jill's v4 file
jillFile4 <- 'C:/Users/deinesji/Google Drive/GEE_validation/valid4_midRep_10_2010_jmd.csv'
# Jill's v3 file
jillFile3 <- 'C:/Users/deinesji/Google Drive/GEE_validation/valid3_midRep_10_2010_jmdTest.csv'

# load/process csvs to spatial points
v4Points <- csvToSpdf(jillFile4, cdlKey)
v3Points <- csvToSpdf(jillFile3, cdlKey)

# project to meters
v4Points <- spTransform(v4Points, CRS('+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=37.5 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs '))
v3Points <- spTransform(v3Points, CRS('+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=37.5 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs '))

# find distances between all points
distPairs <- as.data.frame(gDistance(v4Points, v3Points, byid = T))
distPairs$cellnum <- 1:nrow(distPairs)

# best guess at paired points
pointBuddies <- data.frame(v4 = 1:nrow(v4Points), v3 = NA, distance = NA,
                           crop = v4Points$CLASS_NAME,
                           v4class = v4Points$class, 
                           v3class = NA,
                           v4cert = v4Points$certainty,
                           v3cert = NA)

for (i in 1:nrow(v4Points)) {
  pointBuddies[i,'v3'] <- distPairs[which.min(distPairs[,i]),'cellnum']
  pointBuddies[i, 'distance'] <- distPairs[which.min(distPairs[,i]),i]
}
  
# add irrigation status (class) and certainty for rapp
v3Points$ind <- 1:nrow(v3Points)
pointBuds <- merge(pointBuddies, v3Points[,c('class','certainty','ind')],
                   by.x = 'v3', by.y='ind', all.y = F)
pointBuds$v3class <- pointBuds$class
pointBuds$v3cert <- pointBuds$certainty

# filter for reasonable distances
pointBuds <- pointBuds[pointBuds$distance < 200,]

# summarize agreement
totalMatching <- sum(pointBuds$v3class == pointBuds$v4class)
# percent matching
totalMatching/nrow(pointBuds)

# tabulate irrigated/non-irrigated by crop
pointBuds$identical <- pointBuds$v3class == pointBuds$v4class
table(pointBuds[,c('crop','identical')])
# tabulate by class
table(pointBuds[,c('v4class','identical')])

# check just for certainty level 3
certainPoints <- pointBuds[pointBuds$v4cert == 1 & pointBuds$v3cert ==3,]
matching <- sum(certainPoints$v3class == certainPoints$v4class)
matching/nrow(certainPoints)
# tab by crop
certainPoints$identical <- certainPoints$v3class == certainPoints$v4class
table(certainPoints[,c('crop','identical')])
```

## csv to KML: final versions, valid4
These validations points will be used for further classification

### MidRep 2010

```{r makeKmlValidPts_midrep_2010}
# crop key
cdlKey <- read.csv('C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/helper/CDL_key_2014.csv', stringsAsFactors=F)

fileDir <- 'C:/Users/deinesji/Google Drive/GEE_validation/'

# Jill's valid4 file - Middle Republican NRD
jillFile4 <- paste0(fileDir, 'valid4_midRep_10_2010_jmd.csv')

# load/process csvs to spatial points
v4Points <- csvToSpdf(jillFile4, cdlKey)

# remove system.index column
v4Points <- v4Points[,c('CDLcode','class','classNum','certainty','CLASS_NAME')]

# remove uncertain points
sum(v4Points$certainty == 0) # count number of uncertain poitns = 864
v4Points <- v4Points[v4Points$certainty != 0,]

# write out
#writeOGR(v4Points, paste0(fileDir, 'valid4_midRep_10_2010_jmd_uncertainRemoved.kml'),
#            layer = 'validPoints', driver = 'KML')

# write out only the "most certain" points
v4Points.certain <- v4Points[v4Points$certainty == 1,]
nrow(v4Points.certain)
#writeOGR(v4Points.certain, paste0(fileDir, 'valid4_midRep_10_2010_jmd_certainOnly.kml'), layer = 'validPoints', driver = 'KML')

# write out Shapefiles
outDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/validation/generatedPointsForUpload/shapefiles'
#writeOGR(v4Points, outDir, 'valid4_midRep_10_2010_jmd_uncertainRemoved', driver = 'ESRI Shapefile')
#writeOGR(v4Points.certain, outDir, 'valid4_midRep_10_2010_jmd_certainOnly', driver = 'ESRI Shapefile')

# summarize most certain points by crop
table(v4Points.certain$CLASS_NAME)
table(v4Points.certain$class)
#table(v4Points.certain[,c('class','CLASS_NAME')])
```


### MidRep 2007

```{r makeKmlValidPts_midrep_2007}
# crop key
cdlKey <- read.csv('C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/helper/CDL_key_2014.csv', stringsAsFactors=F)

fileDir <- 'C:/Users/deinesji/Google Drive/GEE_validation/'

# Jill's valid4 file - Middle Republican NRD
jillFile4 <- paste0(fileDir, 'valid4_midRep_10_2007_jmd.csv')

# load/process csvs to spatial points
v4Points <- csvToSpdf(jillFile4, cdlKey)

# remove system.index column
v4Points <- v4Points[,c('CDLcode','class','classNum','certainty','CLASS_NAME')]

# remove uncertain points
sum(v4Points$certainty == 0) # count number of uncertain poitns = 864
v4Points <- v4Points[v4Points$certainty != 0,]

# write out
#writeOGR(v4Points, paste0(fileDir, 'valid4_midRep_10_2007_jmd_uncertainRemoved.kml'), layer = 'validPoints', driver = 'KML')

# write out only the "most certain" points
v4Points.certain <- v4Points[v4Points$certainty == 1,]
nrow(v4Points.certain)
#writeOGR(v4Points.certain, paste0(fileDir, 'valid4_midRep_10_2007_jmd_certainOnly.kml'),layer = 'validPoints', driver = 'KML')

# write out Shapefiles
outDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/validation/generatedPointsForUpload/shapefiles'
#writeOGR(v4Points, outDir, 'valid4_midRep_10_2007_jmd_uncertainRemoved', driver = 'ESRI Shapefile')
#writeOGR(v4Points.certain, outDir, 'valid4_midRep_10_2007_jmd_certainOnly', driver = 'ESRI Shapefile')

# summarize most certain points by crop
v4Points.certain$class <- as.character(v4Points.certain$class)  # unfactor
table(v4Points.certain$CLASS_NAME)
table(v4Points.certain$class)
#table(v4Points.certain[,c('class','CLASS_NAME')])

# all points including uncertain
v4Points$class <- as.character(v4Points$class)  # unfactor
table(v4Points$CLASS_NAME)
table(v4Points$class)
#table(v4Points.certain[,c('class','CLASS_NAME')])
```

### HPA Extra 2010
These are randomly placed 2010 HPA points to enhance the preliminary maps made for Jeremy on 10/30/2016.

```{r makeKmlValidPts_HPAextra_2010}
# crop key
cdlKey <- read.csv('C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/helper/CDL_key_2014.csv', stringsAsFactors=F)

fileDir <- 'C:/Users/deinesji/Google Drive/GEE_validation/'

# Jill's valid4 file - Middle Republican NRD
jillFile4 <- paste0(fileDir, 'valid4_hpa_extra_2010_jmd.csv')

# load/process csvs to spatial points
v4Points <- csvToSpdf(jillFile4, cdlKey)

# remove system.index column
v4Points <- v4Points[,c('CDLcode','class','classNum','certainty','CLASS_NAME')]

# write out
#writeOGR(v4Points, paste0(fileDir, 'valid4_hpa_extra_2010_jmd.kml'), layer = 'validPoints', driver = 'KML')


# all points including uncertain
v4Points$class <- as.character(v4Points$class)  # unfactor
table(v4Points$CLASS_NAME)
table(v4Points$class)
#table(v4Points.certain[,c('class','CLASS_NAME')])
```
