==> 2.00_NASS_countyStats.Rmd <==
---
title: 'NASS irrigation: 2010'
author: "Jill Deines"
date: "Monday, June 20, 2016"
output: 
  html_document:
    toc: yes
---

Goal: Access USDA NASS Ag Census Data through the Quick Stats API, and identify all crops irrigated within boundaries for my remote sensing study.

Note: updated figure used in the SI for the counties 100% within the study area is produced in 2.14_NASS_rsVAlidation_test5randFor.Rmd


```{r knitrOpts, echo=FALSE}
library(knitr)
opts_chunk$set(cache=TRUE, cache.path='cache/2.0_NASS/',
               fig.path='figure/2.0_NASS')
```

**R Packages Needed**

* `httr`: Vignette at https://cran.r-project.org/web/packages/httr/vignettes/quickstart.html

```{r packages, message=FALSE, echo=TRUE}
# for API interaction
library(httr) 
library(jsonlite)
# for county selection:
library(rgdal)
library(rgeos)
library(maps)
# for data manipulation
library(tidyr)
library(ggplot2)

source('functions/getNASS.R')
```

## Functions: NASS Quick Stats API 
This script builds on work in 4.1_USDA_NASS.Rmd in the RRB_Rcode r project.

* API website: http://quickstats.nass.usda.gov/api
* My API key: 28EAA9E6-8060-34A4-981A-B2ED4692228A

Here, I'll make functions to ID counties of interest and pull NASS data from the API, and apply it over the counties.

### Function: Get Counties of Interest
This uses the TIGER county shapefile for the US, extracting counties of interest based on a polygon boundary.

**To do:** deal with counties not entirely inside boundary. 


```{r fun_idCounties}
# # function inputs ------------------------------------
# # load TIGER counties
# countyAll <- readOGR('S:/Data/GIS_Data/Downloaded/TIGER/TIGER2012/COUNTY',
#                      'tl_2012_us_county', verbose=F)
# 
# #polygon denoting area of interest
# gisDir <- 'S:/Users/deinesji/HPA/gis'
# nrds <- readOGR(gisDir,'RRB_NRDs_clip', verbose=F) 
# aoi <- nrds[nrds$NRD_Name == 'Middle Republican',]
# 
# # do a slight negative buffer to avoid line mismatches
# aoi <- gBuffer(aoi, width=-100)
  

# function --------------------------------------------

makeCountyList <- function(countyAll, aoi){
  # match projections
  aoi <- spTransform(aoi, CRS(proj4string(countyAll)))
  
  # id counties inside aoi with a 1. Note `rgeos` must be loaded
  overlap <- gIntersects(countyAll, aoi, byid=T)  # 1's indicate overlap
  
  # remove extra columns and subset overlapping counties
  columnsToKeep <- c('STATEFP','COUNTYFP','COUNTYNS','NAME')
  counties <- countyAll@data[overlap,columnsToKeep]
  
  # fill in state abbrevs column based on FIPS
  data(state.fips)  # from maps package
  state.fips$STATEFP <- sprintf("%02d",state.fips$fips) # add leading zero
  counties <- merge(counties, state.fips[,c('abb','STATEFP')], all.y=F)
  
  return(counties)
}

#test <- makeCountyList(countyAll, aoi)
```

### Function: Extract NASS data by county
This function currently operates at the GROUP level; RRB_Rcode version is at the commodity level (below group); could also do the sector level if desired.

Arguments:

* apiKey: Obtained from http://quickstats.nass.usda.gov/api
* program: 'CENSUS' or 'SURVEY'
* sector: 'CROPS' here (or DEMOGRAPHICS, ECONOMICS, ENVIRONMENTAL, ANIMALS & pRODUCTS)
* group: 'FIELD CROPS' (or HORTICULTURE, VEGETABLES, FRUIT & TREE NUTS)
* aggregation: specify geographic granularity of the data {'STATE','AG DISTRICT','COUNTY','REGION','ZIP CODE'}. Function currently set up for county level only
* state: state 2 character alpha code
* county: county 3-digit ANSI code

**REPLACEd WITH SOURCED 'GETNASS' FUNCTION**


### Function: Clean Field Crop Data
Input: A raw NASS data frame for the 'field crop' group and summary year desired

The function cleans the data and aggregates all counties into total commodities for the input region, split by irrigation status. For Hay, it tallies alfalfa hay/haylage as one category, and "other hay" for all other hay/haylages (after removing superset categories).

Returns a sorted data frame with commodity, irrigated area (km^2), non-irrigated area (km^2), and total area (km^2)

** REPLACE WITH SOURCED FUNCTION?** 

```{r fun_nassFieldCropClean}
# # load data
# outdir <- 'S:/Users/deinesji/HPA/data/USDA_NASS'
# outfile <- 'NASS_FieldCrops_County_midRep_RRB_raw.csv'
# nass.df <- read.csv(paste0(outdir, '/', outfile), stringsAsFactors=F)
# 
# year <- 2012

cleanNassCensus <- function(nass.df, year){
  # remove excess columns
  keepThese <- c('state_alpha','prodn_practice_desc','short_desc','Value',
                 'county_name','unit_desc','commodity_desc','year')
  nass.df <- nass.df[,keepThese]
  
  # remove (D) values, remove commas (!), convert to numeric
  nass.df <- nass.df[!grepl('(D)',nass.df$Value),]
  nass.df$Value <- gsub(",", "", nass.df$Value, fixed = T)
  nass.df$Value <- as.numeric(nass.df$Value)
  
  # keep only records with "ACRES HARVESTED"
  phrase <- 'ACRES HARVESTED'
  nass.df <- nass.df[grepl(phrase,nass.df$short_desc),]
  
  # for now, just 2012 (could split into list by year in future..)
  nass.df <- nass.df[nass.df$year == year,]
  
  # aggregate over counties
  commodityTally <- aggregate(Value ~ short_desc, data = nass.df, FUN = 'sum')
  
  # add back in some helpful columns
  mergeTable <- nass.df[,c('commodity_desc','unit_desc',
                           'prodn_practice_desc','short_desc')]
  mergeTable2 <- unique(mergeTable)
  nass.df2 <- merge(commodityTally, mergeTable2, by.x = 'short_desc', all=F,
                    by.y = 'short_desc',sort=F)
  
  # change 'all production practices' to 'total
  nass.df2[nass.df2$prodn_practice_desc == 'ALL PRODUCTION PRACTICES',
           'prodn_practice_desc'] <- 'TOTAL'
  
  # HAY: make an alfalfa category
  alfalfaRecords <- grepl('ALFALFA', nass.df2$short_desc) & 
                      !grepl('EXCL', nass.df2$short_desc)
  nass.df2[alfalfaRecords,'commodity_desc'] <- 'ALFALFA'
  
  # HAY: delete the supersets to leave just "other" classes
  supersets <- c('HAY & HAYLAGE - ACRES HARVESTED',
                 'HAY & HAYLAGE, IRRIGATED - ACRES HARVESTED',
                 'HAY - ACRES HARVESTED',
                 'HAY, IRRIGATED - ACRES HARVESTED',
                 'HAYLAGE - ACRES HARVESTED',
                 'HAYLAGE, IRRIGATED - ACRES HARVESTED')
  
  nass.df3 <- nass.df2[!nass.df2$short_desc %in% supersets,]
  
  # rename HAY and HAYLAGE to TOTAL HAY
  nass.df3[grepl("HAY",nass.df3$commodity_desc),'commodity_desc'] <- 'OTHER HAY'
  nass.df3[grepl("HAYLAGE",nass.df3$commodity_desc),'commodity_desc'] <- 'OTHER HAY'
  
  # aggregate by crop type and irrigation status
  summary <- aggregate(Value ~ commodity_desc + prodn_practice_desc,
                       data = nass.df3, FUN='sum')
  
  # convert from acres to square kilometers
  summary$Value <- summary$Value / 247.105
  
  # fill in a non-irrigated value
  summary.wide <- spread(summary, key = prodn_practice_desc, value = Value)
  summary.wide[is.na(summary.wide$IRRIGATED),'IRRIGATED'] <- 0
  summary.wide$NON_IRR <- summary.wide$TOTAL - summary.wide$IRRIGATED
  
  # sort
  summary.wide <- summary.wide[rev(order(summary.wide$IRRIGATED)),]
  
  return(summary.wide)
}

#test <- cleanNassCensus(nass.df, year)

```


## Get Data!

### Middle Republican NRD

#### ID counties
Get list of counties from shapefile

```{r getcounties}
# load TIGER counties
countyAll <- readOGR('S:/Data/GIS_Data/Downloaded/TIGER/TIGER2012/COUNTY',
                     'tl_2012_us_county', verbose=F)

#polygon denoting area of interest
gisDir <- 'S:/Users/deinesji/HPA/gis'
nrds <- readOGR(gisDir,'RRB_NRDs_clip', verbose=F) 
aoi <- nrds[nrds$NRD_Name == 'Middle Republican',]

# do a slight negative buffer to avoid line mismatches
aoi <- gBuffer(aoi, width=-100)

# get counties in AOI and print 
counties <- makeCountyList(countyAll, aoi)

counties
```

#### download data
Currently downloads all data for the "field crops" group (all years)

```{r getData_midRepNRD, eval=FALSE}
# set up variables for data calls
apiKey <- '28EAA9E6-8060-34A4-981A-B2ED4692228A'
program = 'CENSUS'
sector <- 'CROPS'
group <- 'FIELD CROPS'
aggregation <- 'COUNTY'

# make empty list to populate, and loop through counties
nass.list <- list()
for(i in 1:nrow(counties)){
  nass.list[[i]] <- getNASS(apiKey, program, sector, group, aggregation, 
                              state = counties$abb[i],
                              county = counties$COUNTYFP[i])
}

# convert list of data frames to 1 giant dataframe
nass.df <- do.call("rbind",nass.list)

# write out raw data
outdir <- 'S:/Users/deinesji/HPA/data/USDA_NASS'
outfile <- 'NASS_FieldCrops_County_midRep_RRB_raw.csv'
write.csv(nass.df, paste0(outdir, '/', outfile), row.names=F)
```

#### Clean Data: Get Irrigated Areas
Loading saved data so not reliant on functioning API. Goal: what crops are irrigated, and in what proportion?

```{r midRep_irrig, fig.height=4, fig.width=8}
# load data
outdir <- 'S:/Users/deinesji/HPA/data/USDA_NASS'
outfile <- 'NASS_FieldCrops_County_midRep_RRB_raw.csv'
nass.df <- read.csv(paste0(outdir, '/', outfile), stringsAsFactors=F)

# clean/summarize data
midRep2012 <- cleanNassCensus(nass.df, year = 2012)
midRep2007 <- cleanNassCensus(nass.df, year = 2007)
#midRep2002 <- cleanNassCensus(nass.df, year = 2002)

# make  a table
kable(midRep2012[,1:3])

# make a plot of irrigated crops
summary.wide2012 <- midRep2012[midRep2012$IRRIGATED > 0,
                              c('commodity_desc','IRRIGATED', 'NON_IRR')]

summary2012 <- gather(summary.wide2012, key = irrig_type, value = Value,
                   IRRIGATED:NON_IRR)

ggplot(summary2012, aes(x=commodity_desc, y=Value, fill=irrig_type)) +
  geom_bar(stat='identity') + 
  ylab('Square Kilometers') + xlab('commodity') + 
  ggtitle('Irrigated Crops - Middle Republican NRD - 2012') + theme_bw()

# do 2007 -----------------------------
# make  a table
kable(midRep2007[,1:3])

# make a plot of irrigated crops
summary.wide2007 <- midRep2007[midRep2007$IRRIGATED > 0,
                              c('commodity_desc','IRRIGATED', 'NON_IRR')]

summary2007 <- gather(summary.wide2007, key = irrig_type, value = Value,
                   IRRIGATED:NON_IRR)

ggplot(summary2007, aes(x=commodity_desc, y=Value, fill=irrig_type)) +
  geom_bar(stat='identity') + 
  ylab('Square Kilometers') + xlab('commodity') + 
  ggtitle('Irrigated Crops - Middle Republican NRD - 2007') + theme_bw()
```

```{r midRepYearFacet, fig.height=6.5, fig.width=8}
# faceted plot
summary2012$year <- 2012
summary2007$year <- 2007

summary0712 <- rbind(summary2012, summary2007)

ggplot(summary0712, aes(x=commodity_desc, y=Value, fill=irrig_type)) +
  geom_bar(stat='identity') + facet_wrap(~ year, ncol=1) +
  ylab('Square Kilometers') + xlab('commodity') + 
  ggtitle('Irrigated Crops - Middle Republican NRD') + theme_bw()
```

### Republican River Basin

```{r fullRRBget, fig.height=4, fig.width = 12, eval=FALSE}
# load TIGER counties ----------------------------------------------------------
countyAll <- readOGR('S:/Data/GIS_Data/Downloaded/TIGER/TIGER2012/COUNTY',
                     'tl_2012_us_county', verbose=F)
#polygon denoting area of interest
gisDir <- 'S:/Users/deinesji/HPA/gis'
rrb <- readOGR(gisDir,'RepBasin', verbose=F) 

# get counties in AOI and print 
counties <- makeCountyList(countyAll, rrb)

# set up variables for data calls ----------------------------------------------
apiKey <- '28EAA9E6-8060-34A4-981A-B2ED4692228A'
program = 'CENSUS'
sector <- 'CROPS'
group <- 'FIELD CROPS'
aggregation <- 'COUNTY'

# make empty list to populate, and loop through counties
nass.list <- list()
for(i in 1:nrow(counties)){
  nass.list[[i]] <- getNASS(apiKey, program, sector, group, aggregation, 
                              state = counties$abb[i],
                              county = counties$COUNTYFP[i])
}

# convert list of data frames to 1 giant dataframe
nass.df.rrb <- do.call("rbind",nass.list)


# write out raw data
outdir <- 'S:/Users/deinesji/HPA/data/USDA_NASS'
outfile <- 'NASS_FieldCrops_County_RRB_raw.csv'
write.csv(nass.df.rrb, paste0(outdir, '/', outfile), row.names=F)
```

Load downloaded data and process

```{r RRB_process, fig.height=4, fig.width = 12}
outdir <- 'S:/Users/deinesji/HPA/data/USDA_NASS'
outfile <- 'NASS_FieldCrops_County_RRB_raw.csv'
nass.df.rrb <- read.csv(paste0(outdir, '/', outfile), stringsAsFactors=F)


# clean/summarize data
rrb2012 <- cleanNassCensus(nass.df.rrb, year = 2012)
rrb2007 <- cleanNassCensus(nass.df.rrb, year = 2007)
#midRep2002 <- cleanNassCensus(nass.df, year = 2002)

# make  a table
kable(rrb2012[,1:3])

# make a plot of irrigated crops
summary.wide2012 <- rrb2012[rrb2012$IRRIGATED > 1.1,
                              c('commodity_desc','IRRIGATED', 'NON_IRR')]

summary2012 <- gather(summary.wide2012, key = irrig_type, value = Value,
                   IRRIGATED:NON_IRR)

ggplot(summary2012, aes(x=commodity_desc, y=Value, fill=irrig_type)) +
  geom_bar(stat='identity') + 
  ylab('Square Kilometers') + xlab('commodity') + 
  ggtitle('Irrigated Crops - Republican River Basin - 2012') + theme_bw()

# do 2007 ---------------------
# make  a table
kable(rrb2007[,1:3])

# make a plot of irrigated crops
summary.wide2007 <- rrb2007[rrb2007$IRRIGATED > 2,
                              c('commodity_desc','IRRIGATED', 'NON_IRR')]

summary2007 <- gather(summary.wide2007, key = irrig_type, value = Value,
                   IRRIGATED:NON_IRR)

ggplot(summary2007, aes(x=commodity_desc, y=Value, fill=irrig_type)) +
  geom_bar(stat='identity') + 
  ylab('Square Kilometers') + xlab('commodity') + 
  ggtitle('Irrigated Crops - Republican River Basin - 2007') + theme_bw()
```

```{r rrbYearFacet, fig.height=7.5, fig.width=12}
# faceted plot
summary2012$year <- 2012
summary2007$year <- 2007

summary0712 <- rbind(summary2012, summary2007)

ggplot(summary0712, aes(x=commodity_desc, y=Value, fill=irrig_type)) +
  geom_bar(stat='identity') + facet_wrap(~ year, ncol=1) +
  ylab('Square Kilometers') + xlab('commodity') + 
  ggtitle('Irrigated Crops - Republican River Basin') + theme_bw()
```

### High Plains Aquifer
dial it up a notch!

Re-run on 1/24/2017 to use the GEE remote sensing HPA boundaring (HPA + 30 km buffer + full RRB)

```{r fullHPAget, eval=FALSE, message=FALSE}
# load TIGER counties ----------------------------------------------------------
countyAll <- readOGR('S:/Data/GIS_Data/Downloaded/TIGER/TIGER2012/COUNTY',
                     'tl_2012_us_county', verbose=F)
#polygon denoting area of interest
gisDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/kml/shapefile'
hpa <- readOGR(gisDir,'HPA_30km_RRB_buff', verbose=F) 

# get counties in AOI and print 
counties <- makeCountyList(countyAll, hpa)

# api fails with i = 183; remove county
counties <- counties[!(counties$COUNTYNS == '01266992'),]

# set up variables for data calls ----------------------------------------------
apiKey <- '28EAA9E6-8060-34A4-981A-B2ED4692228A'
program = 'CENSUS'
sector <- 'CROPS'
group <- 'FIELD CROPS'
aggregation <- 'COUNTY'

# make empty list to populate, and loop through counties
nass.list <- list()
for(i in 1:nrow(counties)){
  nass.list[[i]] <- getNASS(apiKey, program, sector, group, aggregation, 
                              state = counties$abb[i],
                              county = counties$COUNTYFP[i])
}

# convert list of data frames to 1 giant dataframe
#lapply(nass.list, dim) - make sure none over 50000...
nass.df <- do.call("rbind",nass.list)


# write out raw data
outdir <- 'S:/Users/deinesji/HPA/data/USDA_NASS'
outdir2 <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/nass'
outfile <- 'NASS_FieldCrops_County_HPA_raw_full_GEE_aoi_hpa_30km_rrb_buff.csv'
write.csv(nass.df, paste0(outdir, '/', outfile), row.names=F)
write.csv(nass.df, paste0(outdir2, '/', outfile), row.names=F)
```

Load downloaded data and process

```{r fullHPA_process, fig.height=4, fig.width = 8}
outdir <- 'S:/Users/deinesji/HPA/data/USDA_NASS'
outfile <- 'NASS_FieldCrops_County_HPA_raw.csv'
nass.df.hpa <- read.csv(paste0(outdir, '/', outfile), stringsAsFactors=F)

# clean/summarize data
hpa2012 <- cleanNassCensus(nass.df.hpa, year = 2012)
hpa2007 <- cleanNassCensus(nass.df.hpa, year = 2007)
#midRep2002 <- cleanNassCensus(nass.df, year = 2002)

# make  a table
kable(hpa2012[,1:3])

# make a plot of irrigated crops
summary.wide2012 <- hpa2012[hpa2012$IRRIGATED > 1000,
                              c('commodity_desc','IRRIGATED', 'NON_IRR')]

summary2012 <- gather(summary.wide2012, key = irrig_type, value = Value,
                   IRRIGATED:NON_IRR)

ggplot(summary2012, aes(x=commodity_desc, y=Value, fill=irrig_type)) +
  geom_bar(stat='identity') + 
  ylab('Square Kilometers') + xlab('commodity') + 
  ggtitle('Irrigated Crops - High Plains Aquifer - 2012') + theme_bw()

# make  a table
kable(hpa2007[,1:3])

# make a plot of irrigated crops
summary.wide2007 <- hpa2007[hpa2007$IRRIGATED > 1000,
                              c('commodity_desc','IRRIGATED', 'NON_IRR')]

summary2007 <- gather(summary.wide2007, key = irrig_type, value = Value,
                   IRRIGATED:NON_IRR)

ggplot(summary2007, aes(x=commodity_desc, y=Value, fill=irrig_type)) +
  geom_bar(stat='identity') + 
  ylab('Square Kilometers') + xlab('commodity') + 
  ggtitle('Irrigated Crops - High Plains Aquifer - 2007') + theme_bw()
```

```{r hpaYearFacet, fig.height=6.5, fig.width=8}
# faceted plot
summary2012$year <- 2012
summary2007$year <- 2007

summary0712 <- rbind(summary2012, summary2007)

ggplot(summary0712, aes(x=commodity_desc, y=Value, fill=irrig_type)) +
  geom_bar(stat='identity') + facet_wrap(~ year, ncol=1) +
  ylab('Square Kilometers') + xlab('commodity') + 
  ggtitle('Major Irrigated Crops - High Plains Aquifer') + theme_bw()
```

#### Other Tests
I'm curious about irrigated areas in other NASS data groups, incluidng FRUITY & TREE NUTS, HORTICULTURE, and VEGETABLES

nO DATA for 'fruit and tree nuts'

others seem minor

```{r nass_other, eval=FALSE}
# set up variables for data calls
apiKey <- '28EAA9E6-8060-34A4-981A-B2ED4692228A'
program = 'CENSUS'
sector <- 'CROPS'
group <- 'VEGETABLES'
aggregation <- 'COUNTY'

# make empty list to populate, and loop through counties
nass.list <- list()
for(i in 1:nrow(counties)){
  nass.list[[i]] <- getNASS(apiKey, program, sector, group, aggregation, 
                              state = counties$abb[i],
                              county = counties$COUNTYFP[i])
}

# convert list of data frames to 1 giant dataframe
nass.df <- do.call("rbind",nass.list)

```
