==> 2.02_NASS_adk_v_jmd_UpdatedNassCleaningFunction.Rmd <==
---
title: 'NASS: Anthony vs Jill'
author: "Jill Deines"
date: "January 20, 2017"
output: 
  html_document:
    toc: yes
---

Goal: Compare irrigated acres per county tabulated via Anthony's and Jill's scripts for NASS data


```{r knitrOpts, echo=FALSE}
library(knitr)
opts_chunk$set(cache=FALSE, cache.path='cache/2.02_NASS_AvJ/',
               fig.path='figure/2.02_NASS_AvJ/')
```

**R Packages Needed**


```{r packages, message=FALSE, echo=TRUE}
library(ggplot2)
library(tidyr)

# Jill's functions
source('functions/cleanNassCensusCounty.R')
```


## Load Data

And process Jill's from raw downloads to informative data

```{r loadData}
# nass data dir
nassdir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/nass'

# Antonio --------------------------------------
adkfile <- 'county_acreages_fromADK.csv'
adk <- read.csv(paste0(nassdir, '/', adkfile))

# pad 5 digits column of geoID
adk$fips5 <- sprintf("%05d",adk$GEOID)
# make a year column
adk$year <- as.numeric(substr(as.character(adk$year), start = 1, stop = 4))
# just keep irrigated data
adk2 <- adk[adk$production_practice == 'IRRIGATED',]
# add dataset id
adk2$scientist <- 'adk'
# keep columns
adk2 <- adk2[,c('fips5','year','acres_frac','scientist')]

# Jill's data ----------------------------------
nassfile <- 'NASS_FieldCrops_County_RRB_raw.csv'
nass.df <- read.csv(paste0(nassdir, '/', nassfile), stringsAsFactors=F)

# clean data
rrb1997 <- cleanNassCensusCounty(nass.df, year = 1997)
rrb2002 <- cleanNassCensusCounty(nass.df, year = 2002)
rrb2007 <- cleanNassCensusCounty(nass.df, year = 2007)
rrb2012 <- cleanNassCensusCounty(nass.df, year = 2012)

# get irrigated acreage by county
rrb1997.irr <- aggregate(IRRIGATED ~ fips5, data=rrb1997, FUN='sum')
rrb2002.irr <- aggregate(IRRIGATED ~ fips5, data=rrb2002, FUN='sum')
rrb2007.irr <- aggregate(IRRIGATED ~ fips5, data=rrb2007, FUN='sum')
rrb2012.irr <- aggregate(IRRIGATED ~ fips5, data=rrb2012, FUN='sum')

# add years
rrb1997.irr$year <- 1997
rrb2002.irr$year <- 2002
rrb2007.irr$year <- 2007
rrb2012.irr$year <- 2012

# combine
jmd <- rbind.data.frame(rrb1997.irr, rrb2002.irr, rrb2007.irr, rrb2012.irr)

# convert back to acres
jmd$acres_frac <- jmd$IRRIGATED * 247.105

# add dataset id
jmd$scientist <- 'jmd'

# keep columns
jmd <- jmd[,c('fips5','year','acres_frac','scientist')]
```

### Crops in Region SI Figure 2

```{r cropSupplementFig, fig.width=6.5, fig.height=4, dpi=600, dev=c('png','pdf')}
#extract crops with some irrigation
summary.wide2007 <- rrb2007[rrb2007$IRRIGATED > 0,
                              c('commodity_desc','IRRIGATED', 'NON_IRR')]

# filter for...5 main crops?
cropsToKeep <- c('CORN','HAY.OTHER','SORGHUM','SOYBEANS','WHEAT')
summary.wide2007 <- summary.wide2007[summary.wide2007$commodity_desc %in% cropsToKeep,]

summary2007 <- gather(summary.wide2007, key = irrig_type, value = Value,
                   IRRIGATED:NON_IRR)

# fix words for plot
irrStatuskey <- data.frame(irrig_type = c('IRRIGATED','NON_IRR'),
                           plotIrrType = c('Irrigated','Dryland'))
commodityKey <- data.frame(commodity_desc = unique(summary2007$commodity_desc),
                           plot_desc = c('Sorghum','Alfalfa & Hay','Wheat',
                                         'Soybeans','Corn'))
summary2007 <- merge(summary2007, irrStatuskey)
summary2007 <- merge(summary2007, commodityKey)                           


# TOP 5 CROPS IN COUNTIES WITHIN STUDY AREA FOR 2007
position <- c('Corn','Wheat','Sorghum','Soybeans','Alfalfa & Hay')

ggplot(summary2007, 
       aes(x=plot_desc, y=Value, fill=plotIrrType)) +
  geom_bar(stat='identity') + 
  ylab(expression(paste('Area (',km^2,')',sep=''))) + xlab('Commodity') + 
  scale_fill_manual(values=c('#1f78b4','#b2df8a'),
                    labels=c('Rainfed','Irrigated')) +
  scale_x_discrete(limits = position) +
  theme_bw() + theme(legend.title=element_blank(),
                     axis.text=element_text(size=10),
                     legend.text=element_text(size=10),
                     axis.title=element_text(size=11),
                     panel.grid.major = element_blank(),
                     panel.grid.minor = element_blank())




```



## Compare data

```{r totalArea}
# make sure working with same counties
unique(adk2$fips5) == unique(jmd$fips5)

# combine
merged <- rbind.data.frame(jmd, adk2)

# sum area by year
areaAgg <- aggregate(acres_frac ~ year + scientist, data = merged, FUN = 'sum')

# plot
ggplot(areaAgg, aes(year,acres_frac)) +
  geom_bar(aes(fill = scientist), position = 'dodge', stat='identity') +
  ylab('acres irrigated') + xlab('year') +
  scale_x_continuous(name = '', breaks = c(1997, 2002,2007,2012), 
                     labels = c(1997,2002,2007,2012)) +
  ggtitle('Total Acres Irrigated, all (34) counties') + theme_bw()
```


## Error finding

### Compare by Commodity
I had Anthony re-export his results at the commodity level to better identify the discrepancy.

```{r commoditycompareLoad}
# nass data dir
nassdir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/nass'

# Antonio --------------------------------------
adkfile2 <- 'county_acreages_by_commodity_and_practice_fromADK.csv'
adk3 <- read.csv(paste0(nassdir, '/', adkfile2))

# pad 5 digits column of geoID
adk3$fips5 <- sprintf("%05d",adk3$GEOID)
# make a year column
adk3$year <- as.numeric(substr(as.character(adk3$year), start = 1, stop = 4))
# just keep irrigated data
adk4 <- adk3[adk3$production_practice == 'IRRIGATED',]
# convert commodities to long data format
adkLong <- gather(adk4, key = 'commodity_desc', value = 'irr_acres', ALFALFA:HAY.OTHER)
# remove all rows with NA
adkLong2 <- adkLong[!is.na(adkLong$irr_acres),]
# add dataset id
adkLong2$scientist <- 'adk'
# keep columns
adk5 <- adkLong2[,c('fips5','year','commodity_desc','irr_acres','scientist')]

# Jill's data ----------------------------------
nassfile <- 'NASS_FieldCrops_County_RRB_raw.csv'
nass.df <- read.csv(paste0(nassdir, '/', nassfile), stringsAsFactors=F)

# clean data: by commodity and irrigation status
rrb1997 <- cleanNassCensusCounty(nass.df, year = 1997)
rrb2002 <- cleanNassCensusCounty(nass.df, year = 2002)
rrb2007 <- cleanNassCensusCounty(nass.df, year = 2007)
rrb2012 <- cleanNassCensusCounty(nass.df, year = 2012)

# add years
rrb1997$year <- 1997
rrb2002$year <- 2002
rrb2007$year <- 2007
rrb2012$year <- 2012

# combine
jmd2 <- rbind.data.frame(rrb1997, rrb2002, rrb2007, rrb2012)

# just keep irrigated data and convert back to acres
# convert back to acres
jmd2$irr_acres <- jmd2$IRRIGATED * 247.105
jmd2$scientist <- 'jmd'

# columns to keep
jmd3 <- jmd2[,c('fips5','year','commodity_desc','irr_acres','scientist')]
```

### compare adk1 vs adk2
My first look by commodities has higher values for adk than jmd, which is opposite of above. So, check anthony's two datasets against each other

Here, checking adk2 and adk5

```{r adkcheck}
# check number of counties
length(unique(adk2$fips5))
length(unique(adk5$fips5))
```

Well, there you go. adk5 must have all overlapping counties...need to subset!

### Commodity barplots by year

```{r commodityCompare, fig.width = 10.5, fig.height = 5}
# subset adk5 counties to match jmd3
adk6 <- adk5[adk5$fips5 %in% jmd3$fips5,]


# combine datasets
comm <- rbind(jmd3, adk6)

# sum acres by commodity type across counties by year
comm2 <- aggregate(irr_acres ~ commodity_desc + scientist + year, 
                   data = comm, FUN = 'sum')

# plot
ggplot(comm2[comm2$year == 1997,], aes(commodity_desc,irr_acres)) +
  geom_bar(aes(fill = scientist), position = 'dodge', stat='identity') +
  ylab('acres irrigated') + xlab('crop') +
  ggtitle('1997: Total Acres Irrigated, all (34) counties') + theme_bw() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1))

ggplot(comm2[comm2$year == 2002,], aes(commodity_desc,irr_acres)) +
  geom_bar(aes(fill = scientist), position = 'dodge', stat='identity') +
  ylab('acres irrigated') + xlab('crop') +
  ggtitle('2002: Total Acres Irrigated, all (34) counties') + theme_bw() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1))

ggplot(comm2[comm2$year == 2007,], aes(commodity_desc,irr_acres)) +
  geom_bar(aes(fill = scientist), position = 'dodge', stat='identity') +
  ylab('acres irrigated') + xlab('crop') +
  ggtitle('2007: Total Acres Irrigated, all (34) counties') + theme_bw() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1))

ggplot(comm2[comm2$year == 2012,], aes(commodity_desc,irr_acres)) +
  geom_bar(aes(fill = scientist), position = 'dodge', stat='identity') +
  ylab('acres irrigated') + xlab('crop') +
  ggtitle('2012: Total Acres Irrigated, all (34) counties') + theme_bw() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1))
```
