==> 4.10_Figures_YearsIrrigated.Rmd <==
---
title: "Figures"
author: "Jill Deines"
date: "Friday, December 02, 2016"
output: 
  html_document:
    toc: yes
---

Goal: Make figures of irrigation frequency over the region

Update 7/5/2017: changed boundary to the union of the RRCA and RRb


```{r knitrOpts, echo=FALSE}
library(knitr)
opts_chunk$set(cache=FALSE, cache.path='cache/4.1_FreqFigs/test5/union/',
               fig.path='figure/4.1_FreqFigs_randFor/to2016/test5/union/')
```

**R Packages Needed**

```{r packages, message=FALSE, echo=TRUE, eval=TRUE}
library(rgdal)
library(raster)
library(grid)
library(colorRamps)
library(fields)
library(RColorBrewer)
library(latticeExtra)
library(viridis)
library(rasterVis)
```

## Irrigation frequency
Exported raster of years irrigated from 06.0_summaryMaps in GEE.

### User variables

```{r userVars}
# dirs and names
rasDir <- 'C:/Users/deinesji/Google Drive/GEE_AIM-RRB/GEE_classification/RRB_test5_randFor_interannualClean_unionBound'
rasRaw <- 'IrrigationFrequencyMap_0and1neg.tif'
rasMasked <- 'IrrigationFrequencyMap_0and1neg_boundMask.tif'
  
# inset sets
rasSmall1.name <- 'IrrigationFrequencyMap_0and1neg_inset.tif'
earlyInset1.name <- 'earliestYearIrrigated_inset.tif'
lateInset1.name <- 'latestYearIrrigated_inset.tif'

# rasSmall2.name <- 'IrrigationFrequencyMap_0and1neg_inset_2.tif'
# earlyInset2.name <- 'earliestYearIrrigated_inset_2.tif'
# lateInset2.name <- 'latestYearIrrigated_inset_2.tif'
# 
# rasSmall3.name <- 'IrrigationFrequencyMap_0and1neg_inset_3.tif'
# earlyInset3.name <- 'earliestYearIrrigated_inset_3.tif'
# lateInset3.name <- 'latestYearIrrigated_inset_3.tif'

# extent shapefile
extentDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/kml/FINAL_BOUNDARIES'
extentName <- 'RRCA_RRB_union_aea'
rrb <- 'RRB_aea'
```

Set some common color scales, etc, among figures

```{r commonPlotVars}
# number of years color ramps
numYears <- 18
#backgroundcolor <- 'gray60'
backgroundcolor <- 'gray32'
backgroundcolor2 <- 'gray50'

# number of years color ramp
matlab <- matlab.like2(numYears)
#matlab <- viridis(numYears+1)

# year color ramp
# red yellow blue for dave
yearramp <- rev(colorRampPalette(brewer.pal(9,'RdYlBu'))(numYears))

##green/pink for me
#yearramp <- rev(colorRampPalette(brewer.pal(9,'PiYG'))(numYears))
```

end user vars

### map data pre-processing - only need to run 1x
First, load raster domain raster and mask, and write out masked version to save time in the future.

```{r frequencyPrep, eval=FALSE}
# load full raster 
freq <- raster(paste0(rasDir,'/',rasRaw))

# load extent 
extent0 <- readOGR(extentDir, extentName, verbose=F)
extent1 <- spTransform(extent0, CRS(proj4string(freq)))

# set -1 to 0 (areas masked by nlcd)
freq[freq == -1] <- 0

# mask by polygon bound
freq2 <- mask(freq, extent1)

# write out masked raster
writeRaster(freq2, paste0(rasDir,'/',rasMasked))
```

### Plots

### histogram
Frequency histogram of number of years irrigated for pixels: get frequency of values

```{r histPrep, eval=FALSE}
# load full raster
freqHist <- raster(paste0(rasDir,'/',rasRaw))
# count by classes
yearCounts <- freq(freqHist)
# takes a while, save output
write.csv(yearCounts, paste0(rasDir,'/test5_randFor_yearMapCounts.csv'),
          row.names=F) 
```

plot it

```{r hist, fig.width = 6, fig.height = 5, eval=FALSE}
# load yearmap counts
yearCounts <- read.csv(paste0(rasDir,'/test5_randFor_yearMapCounts.csv'))
# sort
yearCounts <- yearCounts[order(yearCounts$value),]

# remove 0 and -1 row
yearCounts2 <- yearCounts[!(yearCounts$value == -1 | yearCounts$value == 0),]
rownames(yearCounts2) <- 1:nrow(yearCounts2)


# get a percentage column: of area ever irrigated
yearCounts2$percentOfIrr <- yearCounts2$count/sum(yearCounts2$count) * 100

# side note: what percentage for 16-18 years? (perennial)
sum(yearCounts2$count[15:17])/sum(yearCounts2$count)

# non-perennial
sum(yearCounts2$count[1:14])/sum(yearCounts2$count)

# arange plot parameters
bardata <- yearCounts2$percentOfIrr
colors <- matlab[2:19]

# vert plot
barplot(bardata, col=colors, space=0, las=1, line=-.5)
mtext(side = 2, "Percent of Irrigated Area", line=2, cex=1.1)
mtext(side = 1, "Years Irrigated", line=4, cex=1.1)

# legend
image.plot(z=0:18,add=F,legend.only=T,  col=c(backgroundcolor,matlab[1:18]), 
           horizontal = T,
          smallplot=c(0.08,.902,.13,.18), 
         axis.args=list(at=c(0,3,6,9,12,15,18), 
                          labels=as.character(c(0,3,6,9,12,15,18))))

```

hist without colorscale

```{r hist_noScale, fig.width = 6.5, fig.height = 5, eval=FALSE}

# vert plot
barplot(bardata, col=colors, space=0.1, las=1, line=-.5,
        names.arg=c('2','','4','','6','','8','','10','','12',
                    '','14','','16','','18'))
mtext(side = 2, "Percent of Irrigated Area", line=2, cex=1.1)
mtext(side = 1, "Years Irrigated", line=2, cex=1.1)

```


### Number of Years Irrigated Maps

Number of years irrigated figure

```{r regionLatLong, fig.width = 8, fig.height = 5, dpi=600, dev=c('png','pdf')}
# manually make polygon to highligh inset area
xmin <- -101.1
xmax <- -100.78
ymin <- 40.76
ymax <- 40.98

insetPOly0 <- Polygon(coords = as.matrix(data.frame(
  lon = c(xmin, xmin, xmax, xmax, xmin),
  lat = c(ymin, ymax, ymax, ymin, ymin))),
  hole=F)
insetPOly00 <- Polygons(list(insetPOly0), ID='scale')
insetPoly <- SpatialPolygons(list(insetPOly00), 
                            proj4string = CRS("+proj=longlat +datum=WGS84"))


extent0 <- readOGR(extentDir, extentName, verbose=F)
extent.ll  <- spTransform(extent0, CRS("+proj=longlat +datum=WGS84"))
plot(extent.ll, axes=T, tck=0,las=1)
axis(3, labels=NA)
axis(2, labels=NA)
plot(insetPoly, add=T)
#plot(freq2)
```



```{r freqMap_matlab2_bottom_matlab, fig.width = 7, fig.height = 5, dpi=600, dev=c('png','pdf'), eval=FALSE}
# load full raster
freq2 <- raster(paste0(rasDir,'/',rasMasked))
rrbpoly <- readOGR(extentDir, rrb, verbose=F)
rrca <- readOGR(extentDir, 'RRCA_modified_aea', verbose=F)

# plot scale
scale = list("SpatialPolygonsRescale", layout.scale.bar(), 
  offset = c(-240000,1785000), scale = 100000, fill=c("transparent","black"))
text1 = list("sp.text", c(-235000,1775000), "0", cex=1)
text2 = list("sp.text", c(-140000,1775000), "100 km",cex = 1)
#arrow = list("SpatialPolygonsRescale", layout.north.arrow(type=1), 
#             offset = c(-223000, 1780000), scale = 40000)

#with borders
spplot(freq2, col.regions = c(backgroundcolor,matlab), maxpixels = 2000000,
       sp.layout=list(scale, text1, text2), #, arrow), 
       par.settings=list(axis.line=list(col=NA)), 
       colorkey=F) +
  latticeExtra::layer(sp.polygons(rrbpoly, lty=2, lwd=1.5)) + 
  latticeExtra::layer(sp.polygons(rrca,lwd=1.5)) 

# without borders
spplot(freq2, col.regions = c(backgroundcolor,matlab), maxpixels = 2000000,
       sp.layout=list(scale, text1, text2), #, arrow), 
       par.settings=list(axis.line=list(col=NA)), 
       #scales = list(draw=T),
       colorkey=F)#list(labels=list(cex=1.2),
                  #at = c(0:18))




#levelplot(freq2, col.regions=c(backgroundcolor,matlab), axes=F)

# color ramp
image.plot(z=0:18,add=F,legend.only=T,  col=c(backgroundcolor3,matlab), 
           horizontal = T,
          smallplot=c(0.05,.95,.15,.2), 
         axis.args=list(at=c(0,3,6,9,12,15,18), 
                          labels=as.character(c(0,3,6,9,12,15,18))))


# scale bar legend title
#grid.text('Years Irrigated',x=unit(0.975, "npc"),y=unit(0.5, 'npc'), rot=-90, gp=gpar(fontsize=18))
```

and a lighter gray to help the outlines pop

```{r freqMap_matlab2_bottom_matlab_lightgray, fig.width = 7, fig.height = 5, dpi=600, dev=c('png','pdf'), eval=FALSE}

backgroundcolor2 <- 'gray50'

#with borders
spplot(freq2, col.regions = c(backgroundcolor2,matlab), maxpixels = 2000000,
       sp.layout=list(scale, text1, text2), #, arrow), 
       par.settings=list(axis.line=list(col=NA)), 
       colorkey=F) +
  latticeExtra::layer(sp.polygons(rrbpoly, lty=2, lwd=1.5)) + 
  latticeExtra::layer(sp.polygons(rrca, lwd=1.5)) 
```

And one for use on dark background slides

```{r freqMap_matlab2_GSA, fig.width = 9, fig.height = 5, dpi = 300, dev = c('pdf','png'), cache=FALSE}
freq2 <- raster(paste0(rasDir,'/',rasMasked))
extent0 <- readOGR(extentDir, extentName, verbose=F)
extent1 <- spTransform(extent0, CRS(proj4string(freq2)))

backgroundcolor3 <- 'black'
#backgroundcolor3 <- '#262626'

# plot scale
scale = list("SpatialPolygonsRescale", layout.scale.bar(), 
  offset = c(-240000,1785000), scale = 100000, fill=c("transparent","black"))
text1 = list("sp.text", c(-235000,1775000), "0", cex=1)
text2 = list("sp.text", c(-140000,1775000), "100 km",cex = 1)
#arrow = list("SpatialPolygonsRescale", layout.north.arrow(type=1), 
#             offset = c(-223000, 1780000), scale = 40000)


spplot(freq2, col.regions = c(backgroundcolor3,matlab), maxpixels = 1000000,
       sp.layout=list(scale), #, arrow), 
       par.settings=list(axis.line=list(col=NA)), 
       colorkey=F)# +
  #latticeExtra::layer(sp.polygons(extent1, lwd=1.5, col='#EEECE1')) 


# color ramp
fields::image.plot(z=0:18,add=F,legend.only=T,  col=c('black',matlab), 
           horizontal = T,
          smallplot=c(0.05,.95,.15,.2), 
         axis.args=list(at=c(0,3,6,9,12,15,18), 
                          labels=as.character(c(0,3,6,9,12,15,18))))

```

One with state lines

```{r freqMap2_GSA_states_nobox_nowatershed, fig.width = 9, fig.height = 5, dpi = 300, dev = 'png', cache=FALSE}
freq2 <- raster(paste0(rasDir,'/',rasMasked))
extent0 <- readOGR(extentDir, extentName, verbose=F)
extent1 <- spTransform(extent0, CRS(proj4string(freq2)))

rrbpoly <- readOGR(extentDir, rrb, verbose=F)
rrca <- readOGR(extentDir, 'RRCA_modified_aea', verbose=F)

states <- readOGR('C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/kml/shapefile', 'States', verbose=F)
states <- states[states$STATE_ABBR %in% c('CO','NE','KS'),]
states <- spTransform(states, CRS(proj4string(freq2)))


backgroundcolor3 <- 'black'
#backgroundcolor3 <- '#262626'

# plot scale
scale = list("SpatialPolygonsRescale", layout.scale.bar(), 
  offset = c(-240000,1785000), scale = 100000, fill=c("transparent","black"))
text1 = list("sp.text", c(-235000,1775000), "0", cex=1)
text2 = list("sp.text", c(-140000,1775000), "100 km",cex = 1)
#arrow = list("SpatialPolygonsRescale", layout.north.arrow(type=1), 
#             offset = c(-223000, 1780000), scale = 40000)


spplot(freq2, col.regions = c(backgroundcolor3,matlab), maxpixels = 1000000,
       #sp.layout=list(scale), #, arrow), 
       par.settings=list(axis.line=list(col=NA)), 
       colorkey=F) +
  #latticeExtra::layer(sp.polygons(extent1, lwd=1, col='#EEECE1')) +
  latticeExtra::layer(sp.polygons(extent1, lty=1, lwd=.8,col= '#EEECE1')) + 
  latticeExtra::layer(sp.polygons(states, lwd=1, col= '#EEECE1'))
  

```

Ghetto way to get a scale bar into AI

```{r freqMap_GSA_scaleBar, fig.width = 9, fig.height = 5, dpi = 300, dev = c('pdf','png'), cache=FALSE}


# plot scale
scale = list("SpatialPolygonsRescale", layout.scale.bar(), 
  offset = c(-240000,1785000), scale = 100000, fill=c("transparent","black"))
text1 = list("sp.text", c(-235000,1775000), "0", cex=1)
text2 = list("sp.text", c(-140000,1775000), "100 km",cex = 1)
#arrow = list("SpatialPolygonsRescale", layout.north.arrow(type=1), 
#             offset = c(-223000, 1780000), scale = 40000)


spplot(freq2, col.regions = c(backgroundcolor3,matlab), maxpixels = 1000,
       sp.layout=list(scale, text1, text2), #, arrow), 
       par.settings=list(axis.line=list(col=NA)), 
       colorkey=F)# +
  #latticeExtra::layer(sp.polygons(extent1, lwd=1.5, col='#EEECE1')) 
```
Integrated colorbar on right (retains map frame border)

```{r freqMap_matlab2_right, eval=FALSE, fig.width = 10, fig.height = 5, dpi=96, eval=FALSE}
# plot scale
scale = list("SpatialPolygonsRescale", layout.scale.bar(), 
  offset = c(-225000,1760000), scale = 100000, fill=c("transparent","black"))
text1 = list("sp.text", c(-220000,1750000), "0")
text2 = list("sp.text", c(-125000,1750000), "100 km")


spplot(freq, col.regions = c(backgroundcolor,matlab), maxpixels = 2000000,
       par.settings = list(layout.widths=list(right.padding=4.5)),
       sp.layout=list(scale, text1, text2),
       colorkey=list(labels=list(cex=1.2),
                     at = c(0:18)))


# scale bar legend title
grid.text('Years Irrigated',x=unit(0.975, "npc"),y=unit(0.5, 'npc'), rot=-90, gp=gpar(fontsize=18))
```

#### inset: Years Irrigated Maps

3 location options

```{r inset1_freqMap_matlab, fig.width = 2.5, fig.height = 3.2, dpi=600, dev=c('png','pdf'), eval=FALSE}
# load full raster
freq_inset1 <- raster(paste0(rasDir,'/',rasSmall1.name))
freq_inset1[freq_inset1 == -1] <- 0



# manually make a scale polygon... :(
xmin <- -406500
ymin <- 1983400
scale <- 5000
height <- 1800

text1 = list("sp.text", c(xmin+400,ymin+300), "3 km", cex=.5)

scalePoly <- Polygon(coords = as.matrix(data.frame(
  lon = c(xmin, xmin, xmin+scale, xmin+scale, xmin),
  lat = c(ymin, ymin+height, ymin+height, ymin, ymin))),
  hole=F)
scalePoly2 <- Polygons(list(scalePoly), ID='scale')
scale.sp <- SpatialPolygons(list(scalePoly2), proj4string = CRS(proj4string(freq_inset1)))


# color ramp 
spplot(freq_inset1, col.regions = c(backgroundcolor,matlab), maxpixels = 200000,
       par.settings = list(layout.widths=list(right.padding=4.5)),
      # sp.layout=list(text1),
       colorkey=list(space = 'bottom',
                     cex = 1.2,
                     at=c(0:18), 
                     labels=c('0','','','3','','','6','','','9',
                              '','','12','','','15','','','18'),
                     width = .85,
                     tck = .7)) +
  latticeExtra::layer(sp.polygons(scale.sp, fill='white', col='black'))
```

inset with lighter backgrounds

```{r inset1_freqMap_matlab_lightgray, fig.width = 2.5, fig.height = 3.2, dpi=600, dev=c('png','pdf'), eval=FALSE}

# color ramp 
spplot(freq_inset1, col.regions = c(backgroundcolor2,matlab), maxpixels = 200000,
       par.settings = list(layout.widths=list(right.padding=4.5)),
      # sp.layout=list(text1),
       colorkey=list(space = 'bottom',
                     cex = 1.2,
                     at=c(0:18), 
                     labels=c('0','','','3','','','6','','','9',
                              '','','12','','','15','','','18'),
                     width = .85,
                     tck = .7)) +
  latticeExtra::layer(sp.polygons(scale.sp, fill='white', col='black'))
```

```{r inset2_freqMap, fig.width = 2.5, fig.height = 3.2, dpi=96, cache=FALSE, eval=FALSE}
# load full raster
freq_inset2 <- raster(paste0(rasDir,'/',rasSmall2.name))
freq_inset2[freq_inset2 == -1] <- 0

# color ramp 
spplot(freq_inset2, col.regions = c(backgroundcolor,matlab), maxpixels = 200000,
       par.settings = list(layout.widths=list(right.padding=4.5)),
       #sp.layout=list(scale, text1, text2),
       colorkey=list(space = 'bottom',
                     cex = 1.2,
                     at=c(0:18), 
                     labels=c('0','','','3','','','6','','','9',
                              '','','12','','','15','','','18'),
                     width = .85,
                     tck = .7))
```

```{r inset3_freqMap, eval=FALSE, fig.width = 2.5, fig.height = 3.2, dpi=96, cache=FALSE}
# load full raster
freq_inset3 <- raster(paste0(rasDir,'/',rasSmall3.name))
freq_inset3[freq_inset3 == -1] <- 0

# color ramp 
spplot(freq_inset3, col.regions = c(backgroundcolor,matlab), maxpixels = 200000,
       par.settings = list(layout.widths=list(right.padding=4.5)),
       #sp.layout=list(scale, text1, text2),
       colorkey=list(space = 'bottom',
                     cex = 1.2,
                     at=c(0:18), 
                     labels=c('0','','','3','','','6','','','9',
                              '','','12','','','15','','','18'),
                     width = .85,
                     tck = .7))
```

#### inset: Earliest Year Irrigated maps


```{r inset1_early, fig.width = 2.5, fig.height = 3.2, dpi=600, dev=c('png','pdf'), cache=FALSE}
# load full raster
early_inset1 <- raster(paste0(rasDir,'/',earlyInset1.name))
early_inset1[early_inset1 == 0] <- NA

# color ramp 
spplot(early_inset1, col.regions = c(yearramp), maxpixels = 200000,
       par.settings = list(layout.widths=list(right.padding=4.5),
                           panel.background=list(col=backgroundcolor)),
       #sp.layout=list(scale, text1, text2),
       colorkey=list(space = 'bottom',
                     cex = 1.2,
                     at=c(1999:2016),
                     labels=c('','2000','','','','','2005','','','',
                              '','2010','','','','','2015',''),
                     width = .85,
                     tck = .7)) #+
  #layer(sp.polygons(extent, fill=backgroundcolor), under=T)
```

```{r inset1_early_lightgray, fig.width = 2.5, fig.height = 3.2, dpi=600, dev=c('png','pdf'), cache=FALSE}

# color ramp 
spplot(early_inset1, col.regions = c(yearramp), maxpixels = 200000,
       par.settings = list(layout.widths=list(right.padding=4.5),
                           panel.background=list(col=backgroundcolor2)),
       #sp.layout=list(scale, text1, text2),
       colorkey=list(space = 'bottom',
                     cex = 1.2,
                     at=c(1999:2016),
                     labels=c('','2000','','','','','2005','','','',
                              '','2010','','','','','2015',''),
                     width = .85,
                     tck = .7)) #+
  #layer(sp.polygons(extent, fill=backgroundcolor), under=T)
```


```{r inset1_early_black, fig.width = 3.5, fig.height = 4.2, dpi=300, dev=c('png','pdf'), cache=FALSE}
# load full raster
early_inset1 <- raster(paste0(rasDir,'/',earlyInset1.name))
early_inset1[early_inset1 == 0] <- NA

# color ramp 
spplot(early_inset1, col.regions = c(yearramp), maxpixels = 200000,
       par.settings = list(layout.widths=list(right.padding=4.5),
                           panel.background=list(col=backgroundcolor3)),
       #sp.layout=list(scale, text1, text2),
       colorkey=F) #+
  #layer(sp.polygons(extent, fill=backgroundcolor), under=T)
```

```{r inset2_early, fig.width = 2.5, fig.height = 3.2, dpi=96, cache=FALSE, eval=FALSE}
# load full raster
early_inset2 <- raster(paste0(rasDir,'/',earlyInset2.name))
early_inset2[early_inset2 == 0] <- NA

# color ramp 
spplot(early_inset2, col.regions = c(yearramp), maxpixels = 200000,
       par.settings = list(layout.widths=list(right.padding=4.5),
                           panel.background=list(col=backgroundcolor)),
            colorkey=list(space = 'bottom',
                     cex = 1.2,
                     at=c(1999:2016),
                     labels=c('','2000','','','','','2005','','','',
                              '','2010','','','','','2015',''),
                     width = .85,
                     tck = .7)) #+
```

```{r inset3_early, fig.width = 2.5, fig.height = 3.2, dpi=96, cache=FALSE, eval=FALSE}
# load full raster
early_inset3 <- raster(paste0(rasDir,'/',earlyInset3.name))
early_inset3[early_inset3 == 0] <- NA

# color ramp 
spplot(early_inset3, col.regions = c(yearramp), maxpixels = 200000,
       par.settings = list(layout.widths=list(right.padding=4.5),
                           panel.background=list(col=backgroundcolor)),
             colorkey=list(space = 'bottom',
                     cex = 1.2,
                     at=c(1999:2016),
                     labels=c('','2000','','','','','2005','','','',
                              '','2010','','','','','2015',''),
                     width = .85,
                     tck = .7)) #+
```

#### inset: Latest Year Irrigated maps

3 location options

```{r inset1_late, fig.width = 2.5, fig.height = 3.2, dpi=600, dev=c('png','pdf'), cache=FALSE}
# load full raster
late_inset1 <- raster(paste0(rasDir,'/',lateInset1.name))
late_inset1[late_inset1 == 0] <- NA

# color ramp 
spplot(late_inset1, col.regions = c(yearramp), maxpixels = 200000,
       par.settings = list(layout.widths=list(right.padding=4.5),
                           panel.background=list(col=backgroundcolor)),
             colorkey=list(space = 'bottom',
                     cex = 1.2,
                     at=c(1999:2016),
                     labels=c('','2000','','','','','2005','','','',
                              '','2010','','','','','2015',''),
                     width = .85,
                     tck = .7)) #+
```

```{r inset1_late_lightgray, fig.width = 2.5, fig.height = 3.2, dpi=600, dev=c('png','pdf'), cache=FALSE}
# color ramp 
spplot(late_inset1, col.regions = c(yearramp), maxpixels = 200000,
       par.settings = list(layout.widths=list(right.padding=4.5),
                           panel.background=list(col=backgroundcolor2)),
             colorkey=list(space = 'bottom',
                     cex = 1.2,
                     at=c(1999:2016),
                     labels=c('','2000','','','','','2005','','','',
                              '','2010','','','','','2015',''),
                     width = .85,
                     tck = .7)) #+
```

```{r inset1_late_black, fig.width = 3.5, fig.height = 4.2, dpi=300, dev=c('png','pdf'), cache=FALSE}

# load full raster
late_inset1 <- raster(paste0(rasDir,'/',lateInset1.name))
late_inset1[late_inset1 == 0] <- NA

# color ramp 
spplot(late_inset1, col.regions = c(yearramp), maxpixels = 200000,
       par.settings = list(layout.widths=list(right.padding=4.5),
                           panel.background=list(col=backgroundcolor3)),
             colorkey=F) #+
```


```{r inset2_late, fig.width = 2.5, fig.height = 3.2, dpi=96, cache=FALSE, eval=FALSE}
# load full raster
late_inset2 <- raster(paste0(rasDir,'/',lateInset2.name))
late_inset2[late_inset2 == 0] <- NA

# color ramp 
spplot(late_inset2,col.regions = c(yearramp), maxpixels = 200000,
       par.settings = list(layout.widths=list(right.padding=4.5),
                           panel.background=list(col=backgroundcolor)),
              colorkey=list(space = 'bottom',
                     cex = 1.2,
                     at=c(1999:2016),
                     labels=c('','2000','','','','','2005','','','',
                              '','2010','','','','','2015',''),
                     width = .85,
                     tck = .7)) #+
```

```{r inset3_late, fig.width = 2.5, fig.height = 3.2, dpi=96, cache=FALSE, eval=FALSE}
# load full raster
late_inset3 <- raster(paste0(rasDir,'/',lateInset3.name))
late_inset3[late_inset3 == 0] <- NA

# color ramp 
spplot(late_inset3, col.regions = c(yearramp), maxpixels = 200000,
       par.settings = list(layout.widths=list(right.padding=4.5),
                           panel.background=list(col=backgroundcolor)),
             colorkey=list(space = 'bottom',
                     cex = 1.2,
                     at=c(1999:2016),
                     labels=c('','2000','','','','','2005','','','',
                              '','2010','','','','','2015',''),
                     width = .85,
                     tck = .7)) #+
```
