==> 4.23_Figures_Landsat.Rmd <==
---
title: "Landsat Figure"
author: "Jill Deines"
date: "April 30, 2017"
output: 
  html_document:
    toc: yes
---

Goal: plot info for Landsat observations

Update 7/18/2017: re-did with new observation counts that remove path duplicates. This also changed the structure of the data cleaning.


```{r knitrOpts, echo=FALSE}
library(knitr)
opts_chunk$set(cache=FALSE, fig.path='figure/4.23_Landsat/')
```

**R Packages Needed**

```{r packages, message=FALSE, echo=TRUE, eval=TRUE}
library(ggplot2)
library(tidyr)
library(gridExtra)
library(grid)

library(rgdal)
library(RColorBrewer)
```

## Scene locations
This needs assembly in AI later

```{r sceneLocations, fig.width = 5.5, fig.height=4.5, dpi=600, dev=c('png','pdf')}
# load data
shapeDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/kml/shapefile'

bound <- readOGR(shapeDir, 'BigExtent_RRB_RRCA_buff', verbose=F)
wrs <- readOGR(shapeDir, 'Landsat_overlapping_scenes_wrs2', verbose=F)
wrs <- spTransform(wrs, CRS=proj4string(bound))

# assign colors by landsat path
pal <- brewer.pal(n=9, name = 'YlGnBu')[4:9]
colorKey <- data.frame(PATH = (unique(wrs$PATH)),
                       color = pal)
wrs <- merge(wrs,colorKey)
wrs$color <- as.character(wrs$color)

# plot
plot(wrs, axes=F, ylab = 'Row', font.lab=2, line=2)
plot(bound, add=T, col='#ffffcc')
plot(wrs, add=T, border=wrs$color)
axis(2, at = c(-100000,060000,220000), labels=c(33,32,31), tck=0, col='white',
     las=1, pos=-420000)
axis(3, at = c(-220000,-100000,020000,140000,260000,380000), 
     labels=c(33:28), tck=0, col='white', pos=320000)
title('Path', cex.main=1)
```

And do a mock-up for all of CONUS for my GEE tutorial:

```{r sceneLocations_CONUS, fig.width = 4, fig.height=3, dpi=300, dev=c('png')}
# load data
shapeDir2 <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/kml/shapefile/CONUS_landsatScenes'

conus <- readOGR(shapeDir2, 'States_continental', verbose=F)
wrs2 <- readOGR(shapeDir2, 'Landsat_scenes_conus_wrs2', verbose=F)
wrs2 <- spTransform(wrs2, CRS=proj4string(bound))
conus <- spTransform(conus, CRS=proj4string(bound))

# assign colors by landsat path
pal2 <- brewer.pal(n=9, name = 'YlGnBu')[5:9]
#pal2 = c('tomato2','royalblue2','springgreen4','black','indianred','purple','darkblue')
#pal2 <- bpy.colors((9))
colorKey2 <- data.frame(PATH = (unique(wrs2$PATH)),
                       color = rep(pal2,10)[1:length(unique(wrs2$PATH))])
wrs2 <- merge(wrs2,colorKey2)
wrs2$color <- as.character(wrs2$color)

# plot
par(mar=c(0,0,0,0))
#plot(wrs2, axes=F, ylab = 'Row', font.lab=2, line=-1)
plot(conus, col='gray75', border='white', lwd=2)
plot(wrs2, add=T, lwd=0.8, border=wrs2$color)
```



## Scene Observations

### Data processing
GEE script 01.07_Landsat_Annual_Statistics out a csv for each year; combine them into one dataset

```{r processData}
dataDir <- 'C:/Users/deinesji/Google Drive/GEE_AIM-RRB/GEE_tableExports/Landsat_Annual_Statistics'

# annual file names
annualFiles <- list.files(dataDir, pattern = '*aim-rrb_extent.csv$')

# year to start
yearStart <- 1999

# read in to a single dataframe
annual = as.data.frame(do.call(rbind, lapply(annualFiles, function(x) {
              csv1 <- read.csv(paste0(dataDir,'/',x))
              csv2 <- csv1[, -which(names(csv1) %in%
                                        c('system.index','.geo'))]
              csv3 <- csv2[csv2$year >= yearStart,]
            })))

# rename columns
names(annual) <- c('Min','X1','Maximum','X25','X5','Median','X75', 
                   'X99', 'Images','Mean','Year')

# make df for plot lines in long format
df.long <- gather(annual[,c('Year','Median','Maximum')], value = value,
                  key = stat, 
                  Median:Maximum)

# extract out information into a df for plot ribbons ---------
perc.99 <- annual[,c('Year','X1','X99')]
names(perc.99) <- c('Year','ymin','ymax')
perc.99$Range <- '1-99'

iqr <- annual[,c('Year','X25','X75')]
names(iqr) <- c('Year','ymin','ymax')
iqr$Range <- '25-75'

ribbons2 <- rbind(perc.99, iqr)
```

### observation stats
range and mean counts

```{r pixelStats}
# remove 2012 for a non-2012 count
annual_no2012 <- annual[annual$Year != 2012,]

# minimum
min(annual_no2012$X1)
min(annual$X1)

# maximum
max(annual_no2012$Maximum)

# overall mean
mean(annual$Mean)
mean(annual_no2012$Mean)
```



### Total annual observations
Plot median + IQR, 1-99 percentiles, and maximum

```{r LandsatObservations, fig.width=6, fig.height=3}

p3 <- ggplot(ribbons2) +
  geom_ribbon(data=ribbons2, aes(ymin=ymin, ymax = ymax, x=Year, fill=Range),
              alpha=0.4) +
  geom_line(data=df.long, aes(x=Year, y=value, linetype=stat)) +
  scale_x_continuous(name = '', 
                  breaks = c(2000,2002,2004,2006,2008,2010,2012,2014,2016), 
                 labels = c(2000,2002,2004,2006,2008,2010,2012,2014,2016),
                 limits = c(1998.5,2016.5)) +
  scale_fill_manual(values = c('1-99'='#bdc9e1',
                               "25-75"='#045a8d'),
                    labels=c('1-99','25-75'),
                    name="Percentile \n Range") +
  scale_linetype_manual(values = c("dashed","solid"),
                        name='') +
    xlab('') + ylab('Pixel Observations') +
  theme_bw() + theme(axis.text=element_text(size=10),
                     legend.text=element_text(size=10),
                     axis.title=element_text(size=11),
                     panel.grid.major = element_blank(),
                     panel.grid.minor = element_blank()) +
  guides(linetype = guide_legend(order = 2),
         fill = guide_legend(order=1))
p3


# # with min/max
# p1 <- ggplot(ribbons) +
#   geom_ribbon(data=ribbons, aes(ymin=ymin, ymax = ymax, x=Year, fill=Range),
#               alpha=0.4) +
#   geom_line(data=df, aes(x=Year, y=Median, lty='Median')) +
#   scale_x_continuous(name = '', 
#                   breaks = c(2000,2002,2004,2006,2008,2010,2012,2014,2016), 
#                  labels = c(2000,2002,2004,2006,2008,2010,2012,2014,2016)) +
#   scale_fill_manual(values = c("Min-Max"='yellow','xminmax'='yellow',
#                                '1-99'='seagreen2','x1-99'='seagreen2',
#                                "25-75"='skyblue3')) +
#   scale_linetype('') +
#   xlab('') + ylab('Pixel Observations') +
#   theme_bw()
# p1
```

### Total number of scenes

```{r landsatSceneNumber, fig.width=6, fig.height=2}

satellites <- data.frame(y1 = c(660,695,660),
                         x1 = c(-Inf,1999,2012.5),
                         xend = c(2011.5, Inf, Inf),
                         Sensor = c('TM','ETM+','OLI'))

p2 <- ggplot(data = annual,
       aes(x=Year, y = Images)) +
    geom_bar(stat='identity', fill='gray60', alpha=.7) + 
  geom_segment(data=satellites,
                aes(x=x1, xend = xend, y=y1, yend=y1, color=Sensor),lwd=2) +
  scale_color_manual(values=c('#a6cee3','#1f78b4','#b2df8a')) +
  scale_x_continuous(name = '', 
                  breaks = c(2000,2002,2004,2006,2008,2010,2012,2014,2016), 
                 labels = c(2000,2002,2004,2006,2008,2010,2012,2014,2016),
                 limits = c(1998.5,2016.5)) +
  xlab('') + ylab('Landsat Scenes') +
  theme_bw() + theme(legend.title=element_blank(),
                     axis.text=element_text(size=10),
                     legend.text=element_text(size=10),
                     axis.title=element_text(size=11),
                     panel.grid.major = element_blank(),
                     panel.grid.minor = element_blank())
p2
```

### All together now
figure arranging from https://gist.github.com/tomhopper/faa24797bb44addeba79

```{r landsatCombined, fig.width=6, fig.height = 5, dpi=600, dev=c('png','pdf')}
grid.newpage()
grid.draw(rbind(ggplotGrob(p2), ggplotGrob(p3), size = "last"))
```
