==> 6.00_IrrigationTrends_test5annual.Rmd <==
---
title: "Irrigation Over Time"
author: "Jill Deines"
date: "April 24, 2017"
output: 
  html_document:
    toc: yes
---

Goal: Investigate trends in irrigated area over time - total area, areas gained, areas lost.

Update 4/9/2017
Update 4/25/2017: for test 5
Update 4/27/2017: added trend from MIRAD for RRCA

Derived datasets produced:

* full domain irrigated area is written out in the NassYearTrends chunk 


```{r knitrOpts, echo=FALSE}
library(knitr)
opts_chunk$set(cache=FALSE, 
               fig.path='figure/6.00_irrTrends_test5annual/')
```

**R Packages Needed**

```{r packages, message=FALSE, echo=TRUE}
library(rgdal)
library(rgeos)
library(latticeExtra)
library(tidyverse) # for ggplot2, tidyr, dplyr
library(broom)
library(colorRamps)
source('functions/stat_smooth_func_all.R')
```

## Load Data

### RS estimates of irrigated area
Counts of each class type were tallied for several administrative boundaries in GEE

```{r loadGee}
cellArea <- 30*30 # m^2

regionsDir <- 'C:/Users/deinesji/Google Drive/GEE_tableExports/RRB_test5_regional_county_Stats'

# get filenames
files <- list.files(regionsDir, pattern="*_test5_randFor_interannual_plusAncillary.csv")

# read in to a single dataframe
all1 = as.data.frame(do.call(rbind, lapply(files, function(x) {
              csv1 <- read.csv(paste0(regionsDir,'/',x))
              csv1$year <- as.numeric(substr(x, start=1,stop=4))
              return(csv1)
            })))

# convert cell count to area
all1$irrigated_m2 <- all1$X1 * cellArea
all1$irrigated_km2 <- all1$irrigated_m2 / 1000000

# find total #cells in the rrcaExtent
#all1$totalCells <- all1$X0 + all1$X1 + all1$null
#all1[all1$masterID == 'rrcaRrbBuffExtent','totalCells'][1]
```

### County Trends
run regressions to get change in area over time (slope) per county

```{r countyRegress, fig.width = 3.5, fig.height = 3, dpi=600, dev=c('png','pdf')}
# extract county data
counties0 <- all1[grepl('county', all1$masterID),]
counties0$fips5 <- substr(counties0$masterID, start = 11, stop = 15)

# drop columns
counties <- counties0[,c('masterID','fips5','year','irrigated_m2','irrigated_km2')]

# Co county 08041 has NA's; remove
counties <- counties[complete.cases(counties),]

# run regressions by county
countylms <- counties %>% 
  group_by(fips5) %>% 
  do(model = lm(irrigated_km2 ~ year, data = .))

# tidy up with broom
coeffs <- countylms %>% tidy(model)
coeffs <- coeffs[coeffs$term == 'year',]

# map the year coefficient
# get county bounds
gisDir <- "C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/kml/shapefile"
cnty <- readOGR(gisDir, 'Counties_RRCAextent_Clip', verbose=F)

# get state bounds
gisDir2 <- "C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/kml"
states <- readOGR(paste0(gisDir2,'/RRB_fullExtent_states.kml'), verbose=F)
states <- spTransform(states, CRS(proj4string(cnty)))

# join the regression slopes
countymap <- merge(cnty, coeffs[,c('fips5','estimate','p.value')])

# or just significant trends
countymap[is.na(countymap$p.value),'p.value'] <- 1 # dummy p value no data county
countymap$sigCoeffs <- countymap$estimate
countymap[countymap$p.value >= 0.05, 'sigCoeffs'] <- NA

# plot!
pal <- matlab.like(16)
# spplot(countymap, 'sigCoeffs', col.regions=pal) +
#   #latticeExtra::layer(sp.polygons(states, col='black', lwd=2))

# give a starkly different color for declining counties
origPal <- pal
origPal[1] <- 'black'
spplot(countymap, 'sigCoeffs', col.regions=origPal, colorkey = list(space='bottom')) #+
 # latticeExtra::layer(sp.polygons(states, col='black', lwd=2))

# # a map with all counties regardless of significance
# origPal2 <- origPal
# origPal2[2] <- 'black'
# spplot(countymap, 'estimate', col.regions=origPal2) +
#   latticeExtra::layer(sp.polygons(states, col='black', lwd=2))
```

Which counties have greatest variance over time?

```{r countyVar}
# which counties are most stable?





```


### Novel/Retired Irrigated Area by year

```{r loadYearly}
cellArea <- 30*30 # m^2

yearsDir <- 'C:/Users/deinesji/Google Drive/GEE_tableExports/earliestLatestTable'

# load 
early <- read.csv(paste0(yearsDir, '/', 'test5_randFor_earliest_regionCounts.csv'))
late <- read.csv(paste0(yearsDir, '/', 'test5_randFor_latest_regionCounts.csv'))

# remove county data
early <- early[!grepl('county', early$masterID),]
late <- late[!grepl('county', late$masterID),]

# gather year-county columns into 1 column
early.long <- gather(early, key = year, value = count, X1999:X2015)
late.long <- gather(late, key = year, value = count, X2000:X2016)

# remove junk columns
early.long <- early.long[,c('masterID','year','count')]
late.long <- late.long[,c('masterID','year','count')]

# convert year column to numeric
early.long$year <- as.numeric(substr(early.long$year, 2,5))
late.long$year <- as.numeric(substr(late.long$year, 2,5))

# convert cell count to area
early.long$area_m2 <- early.long$count * cellArea
early.long$area_km2 <- early.long$area_m2 / 1000000

late.long$area_m2 <- late.long$count * cellArea
late.long$area_km2 <- late.long$area_m2 / 1000000
```

### Combine datasets
Make a long format data frame with total irrigated area, novel area each year, lost area each year

```{r combine}
# reformat total area data ------------------
all <- all1[,c('masterID','year','irrigated_m2','irrigated_km2')]
names(all)[3:4] <- c('area_m2', 'area_km2')
all$dataset <- 'total'

# reformat novel/retired data ----------------
# add dataset ID column
early.long$dataset <- 'novel'
late.long$dataset <- 'retired'

columnsWanted <- c('masterID', 'year', 'area_m2', 'area_km2', 'dataset')
# early, drop first years
early2 <- early.long[!(early.long$year %in% c(1999,2000,2001)),
                       columnsWanted]

# late: drop last years
late2 <- late.long[!(late.long$year %in% c(2014,2015,2016)),
                       columnsWanted]

combined <- rbind(all, early2, late2)
```



## Plot Novel/Retired over time

### full study extent
RRB + RRCA + buffer

```{r fullExtent_novelRetired}
ggplot(data = combined[combined$masterID == 'rrcaRrbBuffExtent',],
       aes(x = year, y = area_km2, colour = dataset)) +
    geom_line() + geom_point() +
    ylab(expression(paste("Irrigated Area (", km^2," )"))) +
    xlab('Year') + ggtitle('Full Region') + 
    theme_bw()

ggplot(data = combined[grepl('full', combined$masterID),],
       aes(x = year, y = area_km2, colour = dataset)) + 
    facet_wrap( ~ masterID) +
    geom_line() + geom_point() +
    ylab(expression(paste("Irrigated Area (", km^2," )"))) +
    xlab('Year') + ggtitle('Full Region by State') +
    theme_bw()
```

Dave and Anthony requested a plot with cumulative novel and retired areas

```{r fullExtent_novelRetired_cumulative_nogrid, fig.width = 6.5, dpi=600, dev=c('png','pdf')}
# extract dataset: full study extent, by state
stateExtent <- combined[grepl('full', combined$masterID),]

# make a cumulative sum column by dataset type
stateExtent <- stateExtent %>%
  group_by(masterID, dataset) %>%
  arrange(year) %>%
  mutate(cumArea_km2 = cumsum(area_km2))

# make a column to plot that has raw data for total, cumulative for novel/retired
stateExtent$toPlot <- NA
stateExtent[stateExtent$dataset == 'total','toPlot'] <- stateExtent[stateExtent$dataset == 'total','area_km2']
stateExtent[stateExtent$dataset == 'novel','toPlot'] <- stateExtent[stateExtent$dataset == 'novel','cumArea_km2']
stateExtent[stateExtent$dataset == 'retired','toPlot'] <- stateExtent[stateExtent$dataset == 'retired','cumArea_km2']

# plotty plot
labels <- c(`CO_full_RRCA` = 'Colorado (CO)',
            `KS_full_RRCA` = 'Kansas (KS)',
            `NE_full_RRCA` = 'Nebraska (NE)')

## legend at the top
# ggplot(data = stateExtent,
#        aes(x = year, y = toPlot, colour = dataset)) + 
#     facet_wrap( ~ masterID, labeller = as_labeller(labels)) +
#     geom_line() + geom_point() +
#     ylab(expression(paste("Irrigated Area (", km^2,")"))) +
#     xlab('Year') + #ggtitle('Full Region by State') +
#     scale_color_discrete(breaks=c('total','novel','retired'),
#                          labels=c('total', 'cumulative novel',
#                                   'cumulative retired')) +
#     guides(color=guide_legend(title=NULL)) +
#     theme_bw() +
#     theme(legend.position = 'top') 

ggplot(data = stateExtent,
       aes(x = year, y = toPlot, colour = dataset)) + 
    facet_wrap( ~ masterID, labeller = as_labeller(labels)) +
    geom_line() + geom_point() +
    ylab(expression(paste("Irrigated Area (", km^2,")"))) +
    xlab('Year') + #ggtitle('Full Region by State') +
    scale_color_manual(breaks=c('total','novel','retired'),
                         labels=c('total', 'cumulative novel',
                                  'cumulative retired'),
                         values=c('red','blue','black')) +
    guides(color=guide_legend(title=NULL)) +
    theme_bw() +
    theme(legend.position = c(.85,.5),
          axis.text=element_text(size=10),
                     legend.text=element_text(size=10),
                     legend.title = element_text(size=10),
                     axis.title=element_text(size=11),
           panel.grid.major = element_blank(),
             panel.grid.minor = element_blank())



```

And a plot of just retired/novel fields over time
```{r fullExtent_novelRetired_alone, fig.height = 2.5, fig.width = 6.5}
# remove total from dataset
stateExtent2 <- stateExtent[stateExtent$dataset != 'total',]

# plotty plot
labels <- c(`CO_full_RRCA` = 'Colorado',
            `KS_full_RRCA` = 'Kansas',
            `NE_full_RRCA` = 'Nebraska')

ggplot(data = stateExtent2,
       aes(x = year, y = area_km2, colour = dataset)) + 
    facet_wrap( ~ masterID, labeller = as_labeller(labels)) +
    geom_line() + geom_point() +
    ylab(expression(paste("Irrigated Area (", km^2,")"))) +
    xlab('Year') + #ggtitle('Full Region by State') +
    guides(color=guide_legend(title=NULL)) +
    theme_bw() +
    theme(legend.position = c(0.15,.6)) 
```

#### Numerical breakdowns
Growth as a percentage of irrigated area, net new fields compared to irrigated area increases...

```{r numerology}
stateData <- combined[grepl('full', combined$masterID),]

# by state, growth as a percentage of 1999 area
percentage <- stateData[stateData$dataset == 'total' &
              stateData$year == '2016', c('masterID','area_km2')]
percentage$startArea <- stateData[stateData$dataset == 'total' &
              stateData$year == '1999', 'area_km2']
percentage$changeArea <- percentage$area_km2 - percentage$startArea

# numbers for novel/retired years
percentage$startArea2 <- stateData[stateData$dataset == 'total' &
              stateData$year == '2002', 'area_km2']
percentage$area_km2_2 <- stateData[stateData$dataset == 'total' &
              stateData$year == '2013', 'area_km2']
percentage$changeAreaShort <- percentage$area_km2_2 - percentage$startArea2
  
percentage$increase_perc <- percentage$area_km2/percentage$startArea * 100

# net acreage added for years where we have data for both
totalCum <- stateExtent2 %>% 
  filter(year %in% c(2002:2013)) %>%
  group_by(masterID, dataset) %>% 
  slice(which.max(cumArea_km2)) 
totalCum <- as.data.frame(totalCum)

# net acreage added for a smorgashbord of years (all years_
totalCum2 <- stateExtent2 %>% 
  group_by(masterID, dataset) %>% 
  slice(which.max(cumArea_km2)) 
totalCum2 <- as.data.frame(totalCum2)

# add net gain to percentage table: short time period
percentage$netKm2_short <- c(totalCum[5,6] - totalCum[6,6],  # NE
                       totalCum[3,6] - totalCum[4,6],  # KS
                       totalCum[1,6] - totalCum[2,6])  # CO

# add net gain to percentage table
percentage$netKm2_all <- c(totalCum2[5,6] - totalCum2[6,6],  # NE
                       totalCum2[3,6] - totalCum2[4,6],  # KS
                       totalCum2[1,6] - totalCum2[2,6])  # CO

# percent of area change due to novel fields, all
percentage$netContribution_all <- percentage$netKm2_all / percentage$changeArea

# percent of area change due to novel fields, 2002-2013
percentage$netContribution_short <- percentage$netKm2_short / percentage$changeAreaShort

percentage

# percent of area change due to novel fields, full region, full time
sum(percentage$netKm2_all) / sum(percentage$changeArea)

# percent of area change due to novel fields, full region, 2002-2013
sum(percentage$netKm2_short) / sum(percentage$changeAreaShort)
```


### basin only
rrb basin bound

```{r basin_novelRetired}


ggplot(data = combined[grepl('basin', combined$masterID) &
                     !grepl('NRD', combined$masterID),],
       aes(x = year, y = area_km2, colour = dataset)) + 
    facet_wrap( ~ masterID) +
    geom_line() + geom_point() +
    ylab(expression(paste("Irrigated Area (", km^2," )"))) +
    xlab('Year') + ggtitle('RRB Basin by State') +
    theme_bw()
```


basin level cumulative

```{r basinExtent_novelRetired_cumulative, fig.width = 6.5}
# extract dataset: full study extent, by state
stateBasin <- combined[grepl('basin', combined$masterID) &
                     !grepl('NRD', combined$masterID),]

# make a cumulative sum column by dataset type
stateBasin <- stateBasin %>%
  group_by(masterID, dataset) %>%
  arrange(year) %>%
  mutate(cumArea_km2 = cumsum(area_km2))

# make a column to plot that has raw data for total, cumulative for novel/retired
stateBasin$toPlot <- NA
stateBasin[stateBasin$dataset == 'total','toPlot'] <- stateBasin[stateBasin$dataset == 'total','area_km2']
stateBasin[stateBasin$dataset == 'novel','toPlot'] <- stateBasin[stateBasin$dataset == 'novel','cumArea_km2']
stateBasin[stateBasin$dataset == 'retired','toPlot'] <- stateBasin[stateBasin$dataset == 'retired','cumArea_km2']

labels <- c(`CO_basin` = 'Colorado',
            `KS_basin` = 'Kansas',
            `NE_basin` = 'Nebraska')

# plotty plot
ggplot(data = stateBasin,
       aes(x = year, y = toPlot, colour = dataset)) + 
    facet_wrap( ~ masterID, labeller = as_labeller(labels)) +
    geom_line() + geom_point() +
    ylab(expression(paste("Irrigated Area (", km^2,")"))) +
    xlab('Year') + #ggtitle('Full Region by State') +
    scale_color_discrete(breaks=c('total','novel','retired'),
                         labels=c('total', 'cumulative novel',
                                  'cumulative retired')) +
    guides(color=guide_legend(title=NULL)) +
    theme_bw() +
    theme(legend.position = 'top') 
```

## Get Trends For Full Study area

### for Year Subsamples (ie, nass)
to demonstrate year sampling affects on understanding of irrigation trends - what would our dataset say if only NASS years?

```{r nassYearTrends}
fullData <- combined[combined$masterID == 'rrcaRrbBuffExtent' &
                       combined$dataset == 'total',]

# write this out to use in Figure 3 (time series)
derivedDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/derivedDatasets'
#write.csv(fullData, paste0(derivedDir, '/FullStudyExtent_irrigation_test5.csv'), 
#          

# overall trend
allYears.lm <- lm(area_km2 ~ year, data = fullData)
summary(allYears.lm)

# NASS years
nassyears <- fullData[fullData$year %in% c(2002,2007,2012),]
nassYears.lm <- lm(area_km2 ~ year, data = nassyears)
summary(nassYears.lm)
```

### full region trends by percent area
Numbers reported in 'Irrigation Trends' section of the manuscript

```{r fullAreaTrend_percent}
# add full study region area
studyArea <- readOGR('C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/kml/shapefile',
'BigExtent_RRB_RRCA_buff', verbose=F)
studyArea <- spTransform(studyArea, CRS('+init=epsg:5070'))
area <- gArea(studyArea)
fullData$studyArea.km2 <- area * 1e-6

# convert percent of total area irrigated
fullData$percentArea <- fullData$area_km2/fullData$studyArea.km2 * 100

# rate of increase, in percent
full.lm <- lm(percentArea ~ year, data = fullData)
summary(full.lm)

# percent increase from low to high
max(fullData$area_km2)/min(fullData$area_km2)

# NASS years
nassyears2 <- fullData[fullData$year %in% c(2002,2007,2012),]
nassYears2.lm <- lm(percentArea ~ year, data = nassyears2)
summary(nassYears2.lm)
```



## Plot Area Data (plots prior to novel/retired analysis)

### Full RRCA RRB Extent

```{r rrcaExtentPlots, fig.height=2.5}
ggplot(data = all1[all1$masterID == 'rrcaRrbBuffExtent',],
       aes(x = year, y = irrigated_km2)) +
    geom_line() + geom_point() +
    ylab(expression(paste("Irrigated Area (", km^2," )"))) +
    xlab('Year') + ggtitle('Full Region') + 
    theme_bw()
# precip
ggplot(data = all1[all1$masterID == 'rrcaRrbBuffExtent',],
       aes(x = year, y=pr_ann)) +
    geom_line() + geom_point() +
    ylab("Dec - Aug precip  (mm)") +
    xlab('Year') + ggtitle('Full Region') + 
    theme_bw()

# aridity
ggplot(data = all1[all1$masterID == 'rrcaRrbBuffExtent',],
       aes(x = year, y=aridity)) +
    geom_line() + geom_point() +
    ylab("Aridity") +
    xlab('Year') + ggtitle('Full Region') + 
    theme_bw()
```

```{r rrcabyState}

ggplot(data = all1[grepl('full', all1$masterID),],
       aes(x = year, y = irrigated_km2, colour = masterID)) +
    geom_line() + geom_point() +
    ylab(expression(paste("Irrigated Area (", km^2," )"))) +
    xlab('Year') + ggtitle('Full Region by State') +
    theme_bw()
```

```{r rrcaStateTrends}
stateRRCA <- all1[grepl('full', all1$masterID),]
stateRRCA$masterID <- as.character(stateRRCA$masterID)

# run regressions by state 
statelms <- stateRRCA %>% 
  group_by(masterID) %>% 
  do(model = lm(irrigated_km2 ~ year, data = .))

# tidy up with broom
coeffs2 <- statelms %>% tidy(model)
coeffs2 <- coeffs2[coeffs2$term == 'year',]

# plots with no facet labels
ggplot(data = stateRRCA,
      aes(x = year, y = irrigated_km2, colour=masterID)) +
  geom_line() + 
  # add linear model
  #stat_smooth_func(geom="text",method="lm",hjust=0,parse=TRUE, 
  #                xpos = 2000, ypos = 14000) +
  # add trendline
  stat_smooth(method = 'lm', se=T) +
  ylab('') +
  theme_bw() + theme(legend.position = 'none',
                     strip.background=element_blank(),
                     strip.text.x = element_blank(),
                     text = element_text(size = 16))

```


### RRB Basin plots

```{r rrbBasinPlots}
ggplot(data = all1[grepl('basin', all1$masterID) &
                     !grepl('NRD', all1$masterID),],
       aes(x = year, y = irrigated_km2, colour = masterID)) +
    geom_line() + geom_point() +
    ylab(expression(paste("Irrigated Area (", km^2," )"))) +
    xlab('Year') + ggtitle('RRB by State') +
   theme_bw() 
```

```{r neBasin, fig.height=2.5}
ggplot(data = all1[grepl('NE_basin', all1$masterID),],
       aes(x = year, y = irrigated_km2)) +
    geom_line() + geom_point() +
    ylab(expression(paste("Irrigated Area (", km^2," )"))) +
    xlab('Year') + ggtitle('Nebraska RRB') +
   theme_bw() 
```
### NRD Basin plots

```{r nrdBasinPlots}
ggplot(data = all1[grepl('NRD', all1$masterID) | 
                     grepl('GMD', all1$masterID),],
       aes(x = year, y = irrigated_km2, colour = masterID)) +
    geom_line() + geom_point() +
    ylab(expression(paste("Irrigated Area (", km^2," )"))) +
    xlab('Year') + ggtitle('By NE NRD') +
   theme_bw() 

```

## MIRAD Trends
What would the trend be if you only had the MIRAD dataset to work from?

```{r mirad}
cellArea <- 250*250 # m^2
rrcaArea <- 79371.68 # km2
studyArea <- 141603 # km2

regionsDir <- 'C:/Users/deinesji/Google Drive/GEE_tableExports/RRB_test5_regional_county_Stats'

# get filenames
files <- list.files(regionsDir, pattern="*_test5_randFor_MIRAD.csv")

# read in to a single dataframe
all2 = as.data.frame(do.call(rbind, lapply(files, function(x) {
              csv1 <- read.csv(paste0(regionsDir,'/',x))
              csv1$year <- as.numeric(substr(x, start=1,stop=4))
              return(csv1)
            })))

# convert cell count to area
all2$irrigated_m2 <- all2$X1 * cellArea
all2$irrigated_km2 <- all2$irrigated_m2 / 1000000 

# extract RRCA bound
rrca.mirad <- all2[all2$masterID == 'RRCA model boundary',]
rrca.mirad$AreaPercentRRCA <- rrca.mirad$irrigated_km2/ rrcaArea * 100

# plot and trend
ggplot(rrca.mirad, aes(x=year, y = AreaPercentRRCA)) +
  geom_line() + geom_point() +
  stat_smooth(method = 'lm', se=F) +
  stat_smooth_func_all(geom="text",method="lm",hjust=0,parse=TRUE, 
                 xpos = 2004) + ggtitle('MIRAD for RRCA')

mirad.lm <- lm(AreaPercentRRCA ~ year, data=rrca.mirad)
summary(mirad.lm)

# extract full study region
full.mirad <- all2[all2$masterID == 'rrcaRrbBuffExtent',]
full.mirad$AreaPercentRRCA <- full.mirad$irrigated_km2/ studyArea * 100

# plot and trend
ggplot(full.mirad, aes(x=year, y = AreaPercentRRCA)) +
  geom_line() + geom_point() +
  stat_smooth(method = 'lm', se=F) +
  stat_smooth_func_all(geom="text",method="lm",hjust=0,parse=TRUE, 
                 xpos = 2004) + ggtitle('MIRAD for Full Area')

mirad2.lm <- lm(AreaPercentRRCA ~ year, data=full.mirad)
summary(mirad2.lm)
```
