==> 6.04_IrrigationTrends_gridsFinalFinal_revisions.Rmd <==
---
title: "Irrigation Over Time"
author: "Jill Deines"
date: "July 5, 2017"
output: 
  html_document:
    toc: yes
---

Goal: Investigate trends in irrigated area over time - total area, areas gained, areas lost. Major updates to 6.00 following GRL review, including using polygon grids instead of counties for trends. 6.03 only includes the final grid size, compared to 6.02 which trialed several.

Derived datasets produced:

* full domain irrigated area is written out in the NassYearTrends chunk 


```{r knitrOpts, echo=FALSE}
library(knitr)
opts_chunk$set(cache=FALSE, cache.path = 'cache/6.04_gridTrends/',
               fig.path='figure/6.04_irrTrends_Final/')
```

**R Packages Needed**

```{r packages, message=FALSE, echo=TRUE,cache=FALSE}
library(rgdal)
library(rgeos)
library(latticeExtra)
library(tidyverse) # for ggplot2, tidyr, dplyr
library(broom)
library(colorRamps)
source('functions/stat_smooth_func_all.R')
library(gridExtra)
library(grid)
```

## Pixel-wise Trends through Time

### load RS estimates of irrigated area
Counts of each class type were tallied for several administrative boundaries in GEE

```{r loadGee}
cellArea <- 30*30 # m^2

regionsDir <- 'C:/Users/deinesji/Google Drive/GEE_AIM-RRB/GEE_tableExports/RRB_test5_regional_county_Stats'

# regional areas -----------------------------------------------
# get filenames
files <- list.files(regionsDir, pattern="*_test5_randFor_interannual_plusAncillary_REVISION_BOUNDARIES_ksSplit.csv")

# read in to a single dataframe
loadStats <- function(fileList){
  datOut = as.data.frame(do.call(rbind, lapply(fileList, function(x) {
              csv1 <- read.csv(paste0(regionsDir,'/',x))
              csv1$year <- as.numeric(substr(x, start=1,stop=4))
              return(csv1)
           })))  
  # convert cell count to area
  datOut$irrigated_m2 <- datOut$X1 * cellArea
  datOut$irrigated_km2 <- datOut$irrigated_m2 / 1000000
  return(datOut)
}

all1 = loadStats(files)

# remove county data 
all1 <- all1[!grepl('county', all1$masterid),]


# find total #cells in the rrcaExtent
#all1$totalCells <- all1$X0 + all1$X1 + all1$null
#all1[all1$masterID == 'rrcaRrbBuffExtent','totalCells'][1]

# gridded areas -----------------------------------------------
# get filenames
files4 <- list.files(regionsDir, pattern="*_test5_randFor_interannual_plusAncillary_POLYGRID_4km.csv")

# read in to a single dataframe, while dropping the "dummy" row
all4 <- loadStats(files4)[-1,]
```

### Gridded Trends
run regressions to get change in area over time (slope) per 4 km grid cell.

```{r gridRegress_fun, cache=TRUE}

# function to get regression coefficients for each cell
getSpatialRegression <- function(dataset, polygrid, p = 0.05, normalize=FALSE){
  # drop columns
  data1 <- dataset[,c('polyid','year','irrigated_m2','irrigated_km2','contained',
                      'area_m2')]
  # remove NAs
  data1 <- data1[!is.na(data1$irrigated_m2),]
  # normalize areal data
  data1$perc_irr <- (data1$irrigated_m2/data1$area_m2) * 100
  
  # run regressions by polygon
  if(normalize == TRUE){
    data1 <- data1[!is.na(data1$perc_irr),]
    lms <- data1 %>% 
      group_by(polyid) %>% 
      do(model = lm(perc_irr ~ year, data = .))
  } else if (normalize == FALSE) {  # use raw data
    lms <- data1 %>% 
    group_by(polyid) %>% 
    do(model = lm(irrigated_km2 ~ year, data = .))
  }
  # tidy up with broom
  coeffs <- lms %>% tidy(model)
  coeffs <- coeffs[coeffs$term == 'year',]
  # merge coefficients with grid polygons
  polygrid$polyid <- as.character(polygrid$polyID)
  coeffs$polyid <- as.character(coeffs$polyid)
  poly <- sp::merge(polygrid, coeffs[,c('polyid','estimate','p.value')])
  # just sig trends
  poly[is.na(poly$p.value),'p.value'] <- 1 # dummy p value no data county
  poly$sigCoeffs <- poly$estimate
  poly[poly$p.value >= p, 'sigCoeffs'] <- NA
  return(poly)
} 

```

Run trends

```{r gridPrep, cache=TRUE}
# load spatial bits
gisDir <- "C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/kml/FINAL_BOUNDARIES"
grid4 <-  readOGR(gisDir, 'Polynet_4km_aea', verbose=F)

# get union bounds
unioned <- readOGR(gisDir, 'RRCA_RRB_union_aea', verbose=F)

# run lms by grid cell: in km^2
poly4 <- getSpatialRegression(all4, grid4, p=0.05, normalize=FALSE)
poly4.union <- raster::intersect(poly4, unioned)

# plot specs

# blue and red color ramps pulled from color brewer (no white)
redRamp2 <- colorRampPalette(c('#fddbc7','#f4a582','#d6604d','#b2182b','#67001f'))
blueRamp2 <- colorRampPalette(rev(c('#d1e5f0','#92c5de','#4393c3','#2166ac','#053061')))

colorbreaks <- 200
biPal <- c(blueRamp2(colorbreaks/2),redRamp2(colorbreaks/2))

# scale range
palMin <- -.6
palMax <- .6 
```

Plot it

```{r gridRegress4_redbluegray_bigger, fig.width = 4.2, fig.height = 2.8, dpi=800, dev=c('png','pdf'), cache=TRUE, eval=FALSE}
# polygon boundaries

rrbpoly <- readOGR(gisDir, 'RRB_aea', verbose=F)
rrca <- readOGR(gisDir, 'RRCA_modified_aea', verbose=F)

states <- readOGR('C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/kml/shapefile', 'States', verbose=F)
states <- states[states$STATE_ABBR %in% c('CO','NE','KS'),]
states <- spTransform(states, CRS(proj4string(rrca)))

# gray background
spplot(poly4.union, 'sigCoeffs', col.regions=biPal, col=NA,
       at=seq(palMin,palMax,(-palMin+palMax)/colorbreaks),
              #main = 'Absolute area, 4 km grid, sig only',
       colorkey=list(space='bottom',
                     labels=list(
                       labels = c('-0.6','-0.4','-0.2','0','0.2','0.4','0.6')))) +
  latticeExtra::layer(sp.polygons(rrbpoly, lty=2, lwd=.8)) + 
  latticeExtra::layer(sp.polygons(rrca,lwd=.8)) +
  latticeExtra::layer(sp.polygons(states, lwd=1, col='gray30')) + 
  latticeExtra::layer(sp.polygons(unioned, fill = 'gray50', col=NA), under=T)
``` 

Plot one for GSA
```{r gridRegress4_redbluegray_GSA_whitelines, fig.width = 5.2, fig.height = 3.8, dpi=300, dev=c('png','pdf'), cache=TRUE, eval=FALSE}
# polygon boundaries

rrbpoly <- readOGR(gisDir, 'RRB_aea', verbose=F)
rrca <- readOGR(gisDir, 'RRCA_modified_aea', verbose=F)

states <- readOGR('C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/kml/shapefile', 'States', verbose=F)
states <- states[states$STATE_ABBR %in% c('CO','NE','KS'),]
states <- spTransform(states, CRS(proj4string(rrca)))

# gray background
spplot(poly4.union, 'sigCoeffs', col.regions=biPal, col=NA,
       at=seq(palMin,palMax,(-palMin+palMax)/colorbreaks),
              #main = 'Absolute area, 4 km grid, sig only',
       colorkey=list(space='bottom',
                     labels=list(
                       labels = c('-0.6','-0.4','-0.2','0','0.2','0.4','0.6')))) +
  latticeExtra::layer(sp.polygons(rrbpoly, lty=2, lwd=1.1, col='#EEECE1')) + 
  latticeExtra::layer(sp.polygons(rrca,lwd=1.1, col='#EEECE1')) +
  latticeExtra::layer(sp.polygons(states, lwd=1.2, col='gray80')) + 
  latticeExtra::layer(sp.polygons(unioned, fill = 'gray20', col=NA), under=T)
``` 

```{r gridRegress4_redbluegray_GSA_black_300_2, fig.width = 4.2, fig.height = 2.8, dpi=800, dev=c('png','pdf'), cache=TRUE}
# polygon boundaries

rrbpoly <- readOGR(gisDir, 'RRB_aea', verbose=F)
rrca <- readOGR(gisDir, 'RRCA_modified_aea', verbose=F)

states <- readOGR('C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/kml/shapefile', 'States', verbose=F)
states <- states[states$STATE_ABBR %in% c('CO','NE','KS'),]
states <- spTransform(states, CRS(proj4string(rrca)))

# gray background
spplot(poly4.union, 'sigCoeffs', col.regions=biPal, col=NA,
       at=seq(palMin,palMax,(-palMin+palMax)/colorbreaks),
              #main = 'Absolute area, 4 km grid, sig only',
       colorkey=list(space='bottom',
                     labels=list(
                       labels = c('-0.6','-0.4','-0.2','0','0.2','0.4','0.6')))) +
  latticeExtra::layer(sp.polygons(rrbpoly, lty=2, lwd=1.1, col='black')) + 
  latticeExtra::layer(sp.polygons(rrca,lwd=1.1, col='black')) +
  latticeExtra::layer(sp.polygons(states, lwd=1.2, col='gray40')) + 
  latticeExtra::layer(sp.polygons(unioned, fill = 'gray60', col=NA), under=T)
``` 

## Novel/Retired Irrigated Area by year
plot total, novel, and retired irrigation by year, by multiple regions of interest

### load data: novel/retired
exported from GEE 03.8_novel_Retired_Irrigation

```{r loadYearly}
cellArea <- 30*30 # m^2

yearsDir <- 'C:/Users/deinesji/Google Drive/GEE_AIM-RRB/GEE_tableExports/earliestLatestTable'

# load 
early <- read.csv(paste0(yearsDir, '/', 'test5_randFor_earliest_revisionBoundaries_ksSplit_regionCounts.csv'))
late <- read.csv(paste0(yearsDir, '/', 'test5_randFor_latest_revisionBoundaries_ksSplit_regionCounts.csv'))

# remove county data
early <- early[!grepl('county', early$masterid),]
late <- late[!grepl('county', late$masterid),]

# gather year-count columns into 1 column
early.long <- gather(early, key = year, value = count, X1999:X2015)
late.long <- gather(late, key = year, value = count, X2000:X2016)

# remove junk columns
early.long <- early.long[,c('masterid','year','count')]
late.long <- late.long[,c('masterid','year','count')]

# convert year column to numeric
early.long$year <- as.numeric(substr(early.long$year, 2,5))
late.long$year <- as.numeric(substr(late.long$year, 2,5))

# convert cell count to area
early.long$area_m2 <- early.long$count * cellArea
early.long$area_km2 <- early.long$area_m2 / 1000000

late.long$area_m2 <- late.long$count * cellArea
late.long$area_km2 <- late.long$area_m2 / 1000000
```

### Combine datasets
Make a long format data frame with total irrigated area, novel area each year, lost area each year

```{r combine}
# reformat total area data ------------------
all <- all1[,c('masterid','year','irrigated_m2','irrigated_km2')]
names(all)[3:4] <- c('area_m2', 'area_km2')
all$dataset <- 'total'

# reformat novel/retired data ----------------
# add dataset ID column
early.long$dataset <- 'novel'
late.long$dataset <- 'retired'

columnsWanted <- c('masterid', 'year', 'area_m2', 'area_km2', 'dataset')
# early, drop first years
early2 <- early.long[!(early.long$year %in% c(1999,2000,2001)),
                       columnsWanted]

# late: drop last years
late2 <- late.long[!(late.long$year %in% c(2014,2015,2016)),
                       columnsWanted]

combined <- rbind(all, early2, late2)
```

### Plot PreProcessing

Extract regions wanted and calculate cumulative added and lost irrigation. Set some overly? complicated plot parameters

```{r plotPrep}
# extract regions to plot
regionsWanted <- c('RRB_KS_lowerBasin','Platte-NE_RRCA','RRB_KS_upperBasin',
                   'RRB_CO_basin','RRB_NE_basin','RRCA_mod_KS_main',
                   'RRCA_RRB_Union','RRCA_modifiedBorders','RRB_basin')

regions0 <- combined[combined$masterid %in% regionsWanted,
                     c('masterid','year','area_km2','dataset')]

# make wide to calculate RRCS KS withouth RRB
regions.wide <- spread(regions0, masterid, area_km2)
regions.wide$KS_non_RRB <- regions.wide$RRCA_mod_KS_main -
                      regions.wide$RRB_KS_upperBasin
# back to long
gathercols <- c('RRB_KS_lowerBasin','Platte-NE_RRCA','RRB_KS_upperBasin',
                   'RRB_CO_basin','RRB_NE_basin','KS_non_RRB',
                   'RRCA_RRB_Union','RRCA_modifiedBorders','RRB_basin',
                'RRCA_mod_KS_main')
regions <- gather_(regions.wide, 'masterid', 'area_km2', 
                   gather_cols=gathercols)
regions <- regions[regions$masterid != 'RRCA_mod_KS_main',]

# make a cumulative sum column by dataset type
regions <- regions %>%
  group_by(masterid, dataset) %>%
  arrange(year) %>%
  mutate(cumArea_km2 = cumsum(area_km2))

# make a column to plot that has raw data for total, cumulative for novel/retired
regions$toPlot <- NA
regions[regions$dataset == 'total','toPlot'] <- regions[regions$dataset == 'total','area_km2']
regions[regions$dataset == 'novel','toPlot'] <- regions[regions$dataset == 'novel','cumArea_km2']
regions[regions$dataset == 'retired','toPlot'] <- regions[regions$dataset == 'retired','cumArea_km2']

# rename masterid's to control order
idKey <- data.frame(masterid = c('RRB_KS_lowerBasin','Platte-NE_RRCA',
                                 'RRB_KS_upperBasin','RRB_CO_basin',
                                 'RRB_NE_basin',
                                 'KS_non_RRB', 'RRCA_RRB_Union',
                                 'RRCA_modifiedBorders', 'RRB_basin'),
                   masterid2 = c('iRRB_KS_lowerBasin', 'fPlatte-NE_RRCA',
                                 'hRRB_KS_upperBasin',
                   'dRRB_CO_basin', 'eRRB_NE_basin', 'gKS_non_RRB',
                   'aRRCA_RRB_Union', 'cRRCA_modifiedBorders', 'bRRB_basin'),
                   rowname = c('three','three','two','two','two','three',
                               'one','one','one'))
regions <- merge(regions, idKey)

# plotty plot
labels <- c(`iRRB_KS_lowerBasin` = 'KS RRB (lower)',
            `fPlatte-NE_RRCA` = 'Platte RRCA',
            `hRRB_KS_upperBasin` = 'KS RRB (upper)',
            `dRRB_CO_basin` = 'CO RRB',
            `eRRB_NE_basin` = 'NE RRB',
            `gKS_non_RRB` = 'KS RRCA',
            `aRRCA_RRB_Union` = 'GRB (all areas)',
            `cRRCA_modifiedBorders` = 'RRCA',
            `bRRB_basin` = 'RRB (full basin)')


# line colors: pull from gridded trend color ramps
redRamp2 <- colorRampPalette(c('#fddbc7','#f4a582','#d6604d','#b2182b','#67001f'))
blueRamp2 <- colorRampPalette(rev(c('#d1e5f0','#92c5de','#4393c3','#2166ac','#053061')))

softer <- c(redRamp2(20)[11],blueRamp2(20)[6],'gray10')
```

### Make Two Plots
9 regions are broken into 2 sets to set y scales different for top row (much larger regions of interest)

Top rows:

```{r novelRetired_largeregions, fig.width = 6.5, fig.height=2.5, dpi=600, dev=c('png','pdf')}
toprow <- ggplot(data = regions[regions$rowname == 'one',],
                 aes(x = year, y = toPlot, colour = dataset)) + 
  facet_wrap( ~ masterid2, labeller = as_labeller(labels)) +
  geom_line() + geom_point() +
  ylab('') +
  xlab('') + 
  scale_color_manual(breaks=c('total','novel','retired'),
                     labels=c('total', 'cumulative novel','cumulative retired'),
                     values=softer) +
  theme_bw() +
  theme(legend.position = 'none',
        strip.background=element_blank(), 
        strip.text.x = element_blank(), 
        axis.text=element_text(size=10),
        legend.text=element_text(size=10),
        legend.title = element_text(size=10),
        axis.title=element_text(size=11),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.background = element_rect(fill = 'transparent', colour = NA),
        plot.background = element_rect(fill = 'transparent', colour = NA),
        axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        text = element_text(size = 11))
toprow
```

bottom rows: 

```{r novelRetired_6regions, fig.width = 6.5, fig.height=4, dpi=600, dev=c('png','pdf')}
botrow <- ggplot(data = regions[regions$rowname != 'one',],
       aes(x = year, y = toPlot, colour = dataset)) + 
    facet_wrap( ~ masterid2, labeller = as_labeller(labels)) +
    geom_line() + geom_point() +
    ylab('') +
    xlab('') + 
    scale_color_manual(breaks=c('total','novel','retired'),
                         labels=c('total', 'cumulative added',
                                  'cumulative removed'),
                         values=softer) +
    guides(color=guide_legend(title=NULL)) +
    theme_bw() +
    theme(legend.position = 'none',
          strip.background=element_blank(), 
          strip.text.x = element_blank(), 
          axis.text= element_text(size=10),
          legend.text=element_text(size=10),
          legend.title = element_text(size=10),
          axis.title=element_text(size=11),
          panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          panel.background = element_rect(fill = 'transparent', colour = NA),
          plot.background = element_rect(fill = 'transparent', colour = NA))
botrow
```

Combo

```{r novelRetired__comboregions, fig.width = 6.5, fig.height = 4.5,dpi=600, dev=c('png','pdf')}
library(gridExtra)
library(grid)
# plot with colors
grid.newpage()
grid.draw(rbind(ggplot2::ggplotGrob(toprow), ggplot2::ggplotGrob(botrow), size = "last"))
```

### create study region map keys
reproducing this here from 0.2 STudy Bounds in order to better integrate with the rest of the figure

```{r loadSubregions, fig.width = 3.25, fig.height = 3, dpi=600, dev=c('png','pdf')}
# load necessary shapefiles
# load spatial bits
gisDir <- "C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/kml/FINAL_BOUNDARIES"
NEplatte <- readOGR(gisDir, 'RRCA-nePlatte-noRRB_aea', verbose=F)
rrbstates3 <- readOGR(gisDir, 'RRB_byState_KS_split_aea', verbose=F)
rrca.rrb.states <- readOGR(gisDir, 'RRCA-RRB_interesection_byState_aea', verbose=F)
rrcastates <- readOGR(gisDir, 'RRCA_modified_byState_aea', verbose=F)

# get union bounds
unioned <- readOGR(gisDir, 'RRCA_RRB_union_aea', verbose=F)
rrb <- readOGR(gisDir, 'RRB_aea', verbose=F)
rrca <- readOGR(gisDir, 'RRCA_modified_aea', verbose=F)

# green orange yellow colors
NEbasin <- '#b2df8a' # light green
platte <- '#33a02c' # darker green

CObasin <- '#ffffcc'#'#ffff99' # light yellow
COall <- 'gold' # darker yellow

KSbasinUP <- '#dfc27d'#'#fdbf6f'#'light orange
KSrrca <- '#8c510a'#'#f46d43'#'dark orange
KSlower <- '#bf812d'#'#fb9a99' # ight red

# crazy plot of main areas  ---------------------
par(mar = c(0,0,0,0))
plot(rrbstates3, border='white')
plot(rrbstates3[rrbstates3$STATE_ABBR == 'NE',], add=T, 
     col = NEbasin)
plot(rrbstates3[rrbstates3$masterID == 'RRB_KS_lowerBasin',], add=T, col=KSlower)
plot(rrcastates[rrcastates$masterID == 'RRCA_mod_KS_main',], add=T, 
     col = KSrrca)
plot(rrca.rrb.states[rrca.rrb.states$masterID == 'RRCARRB_KS_main',], add=T, 
     col=KSbasinUP)
plot(rrcastates[rrcastates$STATE_ABBR == 'CO',], add=T, col=COall)
plot(rrbstates3[rrbstates3$STATE_ABBR == 'CO',], add=T, col = CObasin)
plot(NEplatte, add=T, col=platte)
```

Make a bunch of subregion maps for panel references

co rrb

```{r m_CoRRB, fig.width = 1.8, fig.height = 1.25, dpi=600, dev=c('png','pdf')}
mapFill = 'gray60'
mapHighlight = 'gray85'

par(mar=c(0,0,0,0))
plot(unioned, col=mapHighlight, border=NA)
plot(rrbstates3[rrbstates3$STATE_ABBR == 'CO',], add=T, col = mapFill, border=NA)
```

ne rrb

```{r m_NERRB, fig.width = 1.8, fig.height = 1.25, dpi=600, dev=c('png','pdf')}
par(mar=c(0,0,0,0))
plot(unioned, col=mapHighlight, border=NA)
plot(rrbstates3[rrbstates3$STATE_ABBR == 'NE',], add=T, col = mapFill, border=NA)
```

ks lower

```{r m_KsLower, fig.width = 1.8, fig.height = 1.25, dpi=600, dev=c('png','pdf')}
par(mar=c(0,0,0,0))
plot(unioned, col=mapHighlight, border=NA)
plot(rrbstates3[rrbstates3$masterID == 'RRB_KS_lowerBasin',], add=T, 
     col = mapFill, border=NA)
```

ks upper

```{r m_KsUpper, fig.width = 1.8, fig.height = 1.25, dpi=600, dev=c('png','pdf')}
par(mar=c(0,0,0,0))
plot(unioned, col=mapHighlight, border=NA)
plot(rrca.rrb.states[rrca.rrb.states$masterID == 'RRCARRB_KS_main',], add=T, 
     col = mapFill, border=NA)
```

ks rrca

```{r m_KsRrca, fig.width = 1.8, fig.height = 1.25, dpi=600, dev=c('png','pdf')}
par(mar=c(0,0,0,0))
plot(unioned, col=mapHighlight, border=NA)
plot(rrcastates[rrcastates$masterID == 'RRCA_mod_KS_main',], add=T, 
     col = mapFill, border=NA)
plot(rrca.rrb.states[rrca.rrb.states$masterID == 'RRCARRB_KS_main',], add=T, 
     col = mapHighlight, border=NA)
```

ne platte

```{r m_nePlatte, fig.width = 1.8, fig.height = 1.25, dpi=600, dev=c('png','pdf')}
par(mar=c(0,0,0,0))
plot(unioned, col=mapHighlight, border=NA)
plot(NEplatte, add=T, col = mapFill, border=NA)
```

grb

```{r m_grb, fig.width = 1.8, fig.height = 1.25, dpi=600, dev=c('png','pdf')}
par(mar=c(0,0,0,0))
plot(unioned, col=mapFill, border=NA)
```

rrb

```{r m_rrb, fig.width = 1.8, fig.height = 1.25, dpi=600, dev=c('png','pdf')}
par(mar=c(0,0,0,0))
plot(unioned, col=mapHighlight, border=NA)
plot(rrb, add=T, col = mapFill, border=NA)
```

rrca

```{r m_rrca, fig.width = 1.8, fig.height = 1.25, dpi=600, dev=c('png','pdf')}
par(mar=c(0,0,0,0))
plot(unioned, col=mapHighlight, border=NA)
plot(rrca, add=T, col = mapFill, border=NA)
```

rrb-rrca for Figure 4

```{r m_rrcarrb_fig4, fig.width = 1.8, fig.height = 1.25, dpi=600, dev=c('png','pdf')}
rrbrrca <- readOGR(gisDir, 'RRCA-RRB_interesection_aea', verbose=F)
  
par(mar=c(0,0,0,0))
plot(unioned, col=mapFill, border=NA)
plot(rrbrrca, add=T, col = 'gray32', border=NA)
```

            
## Numerical breakdowns
stats used in the text

Growth as a percentage of irrigated area, net new fields compared to irrigated area increases...

Numbers for boundaries for manuscript revisions

```{r revisedNumerology}
# largest change in the GRB (full study area, RRB union RRCA) - percent increase
union <- all[all$masterid == 'RRCA_RRB_Union',]
max(union$area_km2)/min(union$area_km2)*100-100

# rate of change in union
union$totalArea <- unioned$areakm2[1]
union$percent <- union$area_km2/union$totalArea*100
summary(lm(percent ~ year, data=union))

# trend via Mirad ----------------------
# load data
regionsDir <- 'C:/Users/deinesji/Google Drive/GEE_tableExports/RRB_test5_regional_county_Stats'
files <- list.files(regionsDir, pattern="*_test5_randFor_MIRAD_revisionBounds.csv")
all2 = as.data.frame(do.call(rbind, lapply(files, function(x) {
              csv1 <- read.csv(paste0(regionsDir,'/',x))
              csv1$year <- as.numeric(substr(x, start=1,stop=4))
              return(csv1)
            })))
# convert cell count to area
cellArea <- 250*250 # m^2
all2$irrigated_m2 <- all2$X1 * cellArea
all2$irrigated_km2 <- all2$irrigated_m2 / 1000000 

# extract union  bound
union.mirad <- all2[all2$masterid == 'RRCA_RRB_Union',]
union.mirad$Percent<- union.mirad$irrigated_km2/ unioned$areakm2[1] * 100
# model
mirad.lm <- lm(Percent ~ year, data=union.mirad)
summary(mirad.lm)


```
