==> 6.05_StatisticalModels.Rmd <==
---
title: "Statistical Model"
author: "Jill Deines"
date: "February 27, 2017"
output: 
  html_document:
    toc: yes
---

Goal: Investigate relationship among irrigated area, precip, pumping volume, and crop prices over time 

```{r knitrOpts, echo=FALSE}
library(knitr)
opts_chunk$set(cache=FALSE, 
               fig.path='figure/6.05_statModels/')
```

**R Packages Needed**

```{r packages, message=FALSE, echo=TRUE}
library(ggplot2)
library(GGally)
library(dplyr)
library(tidyr)
library(corrplot)
library(Hmisc)
```

## Data wrangling

### NASS crop price data

First, make a table of inflation adjustments relative to 2015, the final year available in my NASS price dataset. Write out so this is static and not dependent on external download

```{r inflationTable, eval=FALSE}
# adjust for inflation
library(lubridate) 
library(dplyr) 

# download the consumer price index and get adjustments 
monthly_cpi <- read.csv("http://research.stlouisfed.org/fred2/data/CPIAUCSL.csv",
                        header = TRUE)
monthly_cpi$cpi_year <- year(monthly_cpi$DATE)
yearly_cpi <- monthly_cpi %>% group_by(cpi_year) %>% summarize(cpi = mean(VALUE))
yearly_cpi$adj_factor_2015 <- yearly_cpi$cpi/yearly_cpi$cpi[yearly_cpi$cpi_year == 2015]
# for posterity
yearly_cpi$adj_factor_2016 <- yearly_cpi$cpi/yearly_cpi$cpi[yearly_cpi$cpi_year == 2016]
yearly_cpi$adj_factor_2017 <- yearly_cpi$cpi/yearly_cpi$cpi[yearly_cpi$cpi_year == 2017]

# write out
nassdir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/nass/price'
write.csv(yearly_cpi, paste0(nassdir, '/cpi_priceAdjustments.csv'), 
          row.names=F)
```

Price data was downloaded manually from the NASS quickstats tool (https://quickstats.nass.usda.gov/) for the following query: SURVEY - CROPS - FIELD CROPS - CORN/SOYBEANS/WHEAT/HAY/SORGHUM - PRICE RECEIVED - 

* CORN, GRAIN - PRICE RECEIVED, MEASURED IN $ / BU
* HAY, PRICE RECEIVED, MEASURED IN $ / TON (short ton)
* SORGHUM, GRAIN - PRICE RECEIVED, MEASURED IN $ / CWT (hundredweight)
* SOYBEANS - PRICE RECEIVED, MEASURED IN $ / BU
* WHEAT - PRICE RECEIVED, MEASUTED IN $ / BU

- TOTAL - NATIONAL - US TOTAL - 1996:2015 - ANNUAL - YEAR


Here, I load the data and clean it to have consistent pricing across crops and years

```{r nassPrice}
# load nass
nassdir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/nass/price'
price0 <- read.csv(paste0(nassdir, '/nass_prices_5crops_national_1996-2015.csv'))

# load consumer price index adjustments
cpi <- read.csv(paste0(nassdir, '/cpi_priceAdjustments.csv'))

# make a units columns
unitkey <- data.frame(Commodity = c('CORN','HAY','SORGHUM','SOYBEANS','WHEAT'),
                      unit = c('BU','TON','CWT','BU','BU'))
price0 <- merge(price0,unitkey)

# remove silly columns
price0 <- price0[,c('Commodity','Year','Value','unit')]

# add the adjustment factor
price <- merge(price0, cpi[,c('cpi_year','adj_factor_2015')], by.x = 'Year', 
               by.y = 'cpi_year')

# adjust prices (note: matched trials at https://data.bls.gov/cgi-bin/cpicalc.pl)
price$price_2015 <- price$Value / price$adj_factor_2015

# plot it - unadjusted
ggplot(data = price, aes(x=Year, y = Value)) +
  geom_line() +
  facet_wrap(~Commodity, scales='free_y') + theme_bw() +
  ggtitle('Nomial (unadjusted) Crop Prices') 


# plot it
ggplot(data = price, aes(x=Year, y = price_2015)) +
  geom_line() +
  facet_wrap(~Commodity, scales='free_y') + theme_bw() +
  ggtitle('Inflation-adjusted Crop Prices')

# extract only corn
cornprice <- price[price$Commodity == 'CORN',]

ggplot(data = cornprice, aes(x=Year, y = price_2015)) +
  geom_line() + theme_bw() +
  ggtitle('Inflation-adjusted Corn Prices')

# write out dataset
derivedDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/derivedDatasets'
write.csv(cornprice, paste0(derivedDir, '/price_nass_corn_1996-2015.csv'), 
          row.names=F)
```

## Analyses

### Load Data
loads datasets produced/cleaned from other scripts

```{r loadData}
derivedDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/derivedDatasets'

# usda nass annual corn price, $ / bushel
cornprice <- read.csv(paste0(derivedDir, '/price_nass_corn_1996-2015.csv'))

# irrigation area and precipitation
area.state <- read.csv(paste0(derivedDir, '/RRCA_fullModel_area-rfTest3_ppt-dec-aug.csv'))

# irrigation volume
volume.state <- read.csv(paste0(derivedDir, '/RRCA_fullModel_irrigation_volume.csv'))

# merge pumping, area, precip data
byState <- merge(area.state, volume.state, by='key', all.x = T)
dropCols <- c('key','masterID','state.y','year.y')
byState <- byState %>% select(-one_of(dropCols))
names(byState)[c(1,4)] <- c('year','state')

# aggregate to full RRCA model region
rrcairrigation <- aggregate(cbind(irrigated_km2, total_km3, gwirr_km3, swirr_km3, coirr_km3)
                            ~ year,
                            data = byState, FUN = 'sum', na.rm=T)
# that loses 2016 irrigation data
get2016area <- aggregate(irrigated_km2 ~ year, data = byState, FUN = 'sum')
# and average precip
precip <- aggregate(pr_ann ~ year, data = byState, FUN = 'mean')

# merge precip into dataset
rrca0 <- merge(precip,rrcairrigation)

# merge price
rrca <- merge(rrca0, cornprice[,c('Year','price_2015')], by.x = 'year', by.y = 'Year')

# just use total irrigation volume
rrca <- rrca %>% select(-one_of(c('gwirr_km3','swirr_km3','coirr_km3')))
names(rrca) <- c('Year','Precip','Area_km2','Volume_km3','Price')

# add a price lag
rrca$priceLag1 <- cornprice$price_2015[3:19]

```

### Exploratory analyses

Average application depth: Divide irrigated pumping by area (average application depth); plot vs precip expressed in mm 

what proportion of water demand they are satisfying?

```{r applicationDepth, fig.height=3, fig.width=3}
rrca$depth_m <- (rrca$Volume_km3/rrca$Area_km2)*1e3

ggplot(rrca, aes(Precip, depth_m)) +
  geom_point() + theme_bw() +
  ggtitle('Precip vs Application Depth')
```

And plot all time series

```{r timeSeriesPlots, fig.height = 5.5, fig.width = 6}
# adjust names for plot order
rrcaPlot <- rrca
names(rrcaPlot) <- c('Year','zPrecipitation','cArea','dVolume','aPrice','tDepth')
# gather
plotdata <- gather(rrcaPlot, key = variable, value = value,
                          zPrecipitation:tDepth)

# plots with no facet labels
ggplot(data = plotdata,
      aes(x = Year, y = value, colour=variable)) +
  geom_line() + geom_point() +
  scale_colour_manual(values=c('orange','black','red','forestgreen','blue')) + 
  facet_wrap(~variable, nrow=5, scales='free_y') +
  scale_x_continuous(name = '', 
                     breaks = c(2000,2002,2004,2006,2008,2010,2012,2014), 
                     labels = c(2000,2002,2004,2006,2008,2010,2012,2014)) +
  ylab('') +
  theme_bw() + theme(legend.position = 'none',
                     strip.background=element_blank(), # comment to add labels
                     strip.text.x = element_blank(),   # comment to add labels
                     text = element_text(size = 16))

```


Let's look at some preliminary correlations

```{r corrPlots, fig.height = 5, fig.width=5}
# cor plots
varcor <- cor(rrca)
corrplot(varcor, method='color', addCoef.col='black', tl.col='black',
         type = 'lower',
         col = colorRampPalette(c("darkmagenta", "white", "darkgreen"))(200))

ggcorr(rrca, palette = "RdBu", label = TRUE)

# sig levels
corsig <- rcorr(as.matrix(rrca), type = 'pearson')
options(scipen = 999)
corsig$P
```

Interpretation:

* Precip has inverse correlations with irrigation volume and irrigation area
* Volume and area are negatively correlated
* Price Lag 1 correlated with area
* Depth negatively corrected with area, precip; positive with volume

```{r scatterplots, fig.height=5.5, fig.width=8.5}
# an extra function to add trendlines
my_fn <- function(data, mapping, ...){
  p <- ggplot(data = data, mapping = mapping) + 
    geom_point() + 
    geom_smooth(method=loess, fill="red", color="red", ...) +
    geom_smooth(method=lm, fill="blue", color="blue", ...)
  p
}

# scatterplot matrix
ggpairs(rrca, lower=list(continuous = my_fn)) 
```

Hm. A two step model? Predict application depth from precip; predict irrigated area from depth and price? Or a hierarchical model where depth comes from precip...


### Auto-correlation

```{r autocorrelation, fig.height = 5.5, fig.width=8}
# test for autocorrelation in time series
vars <- names(rrca)

# test all vars
par(mfrow=c(3,3))
acfList <- list()
#names(acfList) <- vars
for (var in vars){
  acf(rrca[,var], main = var)
}
```

```{r autocor2, fig.height = 5.5, fig.width = 6}
# plot first differences: calculate
rrca.firstdiff <- as.data.frame(matrix(, nrow = nrow(rrca)-1, ncol=length(vars)))
names(rrca.firstdiff) <- vars
for (var in vars) {
 rrca.firstdiff[,var] <- diff(ts(rrca[,var]), differences=1)
}

rrca.firstdiff$year <- 2000:2015
rrca.firstdiff <- rrca.firstdiff[ , -which(names(rrca.firstdiff) %in% c('Price'))]

# longify
names(rrca.firstdiff) <- c('Year','zPrecipitation','cArea','dVolume','aPrice','tDepth', 'year')
firstdiff <- gather(rrca.firstdiff, key = variable, value = value, zPrecipitation:tDepth)

ggplot(data = firstdiff,
      aes(x = year, y = value, colour=variable)) +
  geom_line() + geom_point() +
  scale_colour_manual(values=c('orange','black','red','forestgreen','blue')) + 
  facet_wrap(~variable, nrow=5, scales='free_y') +
  scale_x_continuous(name = '', 
                     breaks = c(2000,2002,2004,2006,2008,2010,2012,2014), 
                     labels = c(2000,2002,2004,2006,2008,2010,2012,2014)) +
  ylab('') +
  theme_bw() + theme(legend.position = 'none',
                     strip.background=element_blank(), # comment to add labels
                     strip.text.x = element_blank(),   # comment to add labels
                     text = element_text(size = 16))
```

### CaRT
I picked 'ctree' function the party package because of the nice-ish output on an R Blogger's post, but Andy subsequently suggested it as well. Look at that!

```{r cart, fig.height = 5.5, fig.width = 7.5}
library(party)

# control tree
controls <- ctree_control(minsplit = 2, minbucket = 3, mincriterion = .7)

# run tree
ct <- ctree(Area_km2 ~ Precip + Volume_km3 + priceLag1, data = rrca, 
            controls = controls)
plot(ct, main="Conditional Inference Tree")
 

# control tree
controls <- ctree_control(minsplit = 2, minbucket = 2, mincriterion = .7)

# run tree
ct <- ctree(Area_km2 ~ Precip + priceLag1, data = rrca, 
            controls = controls)
plot(ct, main="Conditional Inference Tree")


```
