==> 9.00_GSA_depth_analysis.Rmd <==
---
title: "Irrigation Depth in the RRCA"
author: "Jill Deines"
date: "October 15, 2017"
output:
  html_document:
    toc: yes
---

Goal: Model linear trends in irrigation depth over time in the RRCA

```{r knitrOpts, echo=FALSE}
library(knitr)
opts_chunk$set(cache=FALSE, 
               fig.path='figure/9.0_rrcaDepth/')
```

**R Packages Needed**

```{r packages, message=FALSE, echo=TRUE}
library(rgdal)
library(raster)
library(tidyverse)
library(broom)
library(sf)
library(latticeExtra)
```

## Make RRCA polygon grid
Extract polygon grids of a few resolutions for the RRCA model

* native: 1 mile resolution
* aggregated: 2 mile, 4 mile


```{r rrcaPolygrids, eval=FALSE}
outRas <- 'S:/Users/deinesji/HPA/gis/Wells/PumpingByGridCell/modelGridRasters'

# load a buffered RRCA boundary
buffer <- readOGR('S:/Users/deinesji/HPA/gis/lhm','RRB_bound2_6500m',verbose=F)

# create template raster grid based on RRCA model grid (165 rows, 326 columns)  
CRS <- '+proj=utm +zone=14 +ellps=clrk66 +datum=NAD27 +units=us-ft +no_defs'
template <- raster(crs=CRS, nrows = 165, ncols=326, resolution = 5280,
                 ymn = 14092806, ymx = 14964006, 
                 xmn = 266023, xmx = 1987023)
template[] <- 1:53790 

# mask with boundary 
buffer.utm <- spTransform(buffer, crs(CRS))
ras1mile <- mask(template, buffer.utm)
writeRaster(ras1mile, paste0(outRas,'/rrca_1mile_buffMasked.tif'))
plot(ras1mile, main = 'default 1 mile grid')

# aggregate to 2 and 4mile, assigning lowest cell number in new region
ras2mile <- aggregate(ras1mile, fact = 2, fun = min, expand=FALSE)
ras4mile <- aggregate(ras2mile, fact = 2, fun = min, expand=FALSE)
plot(ras4mile, main = '4 mile grid')

# convert to polygons and write out
poly1mile <- rasterToPolygons(ras1mile)
poly2mile <- rasterToPolygons(ras2mile)
poly4mile <- rasterToPolygons(ras4mile)
plot(poly4mile, main = '4 mile poly grid')

outPolyDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/kml/FINAL_BOUNDARIES'

# writeOGR(poly1mile, outPolyDir, 'RRCA_polygrid_1mile', driver = 'ESRI Shapefile')
# writeOGR(poly2mile, outPolyDir, 'RRCA_polygrid_2mile', driver = 'ESRI Shapefile')
# writeOGR(poly4mile, outPolyDir, 'RRCA_polygrid_4mile', driver = 'ESRI Shapefile')
```

## Load Data
Pumping data is from the RRCA groundwater model and associated files obtained by Xiao from the guy who works on the model (Sam?). Processed data copied over from S:\Users\deinesji\HPA\gis\Wells\PumpingByGridCell, 
which was produced with R script 4.3_Pumping_Data.Rmd in the RRB_Rcode R project over
in Jills s drive. see that folder's readme.txt and/or script for data source and notes

### Irrigation data
And make a total water column

```{r loadpumping, warning=FALSE}
# irr water master file
pumpDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/pumping'
pumpFile <- 'RRCA_irrAllByCell_1999-2015_longFormat_gw_sw_co_FIXED_20171017.csv'
irrByCell0 <- read.csv(paste0(pumpDir, '/',pumpFile))

# set NAs to 0 (warning is for text field, no problem)
colsToZero <- c('swirr_m3','swirr_km3','swirr_acreft','coirr_m3','coirr_km3','coirr_acreft')
irrByCell0[,colsToZero][is.na(irrByCell0[,colsToZero])] <- 0 

# drop extra columns and calculate total irr
irrByCell <- irrByCell0 %>%
  mutate(totalIrr_m3 = gwirr_m3+swirr_m3+coirr_m3,
         unique2 = paste0(Year, '-', NODE)) %>%
  dplyr::select(unique2, NODE, inRRB, state, Year, totalIrr_m3)  %>%
  rename(year = Year)
```

### Load irrigated area by RRCA cell
irrigated pixel counts tallied via GEE script 04.13_RegionalStats_polygrid_interannual in users/jdeines/default/Production

The 1 mile grid gave some bizarre depths, using a larger aggregation (4 mile)

```{r loadAreas}
cellArea <- 30*30 # m^2

areaDir <-'C:/Users/deinesji/Google Drive/GEE_AIM-RRB/GEE_tableExports/RRB_test5_regional_county_Stats'

# regional areas -----------------------------------------------

# function to read in to a single dataframe
loadStats <- function(fileList){
  datOut = as.data.frame(do.call(rbind, lapply(fileList, function(x) {
              csv1 <- read.csv(paste0(areaDir,'/',x)) %>%
                mutate(X1 = replace(X1, is.na(X1),0))
              csv1$year <- as.numeric(substr(x, start=1,stop=4))
              return(csv1)
           })))  
  # convert cell count to area
  datOut$irrigated_m2 <- datOut$X1 * cellArea
  datOut$irrigated_km2 <- datOut$irrigated_m2 / 1000000
  return(datOut)
}

# # load 1 mile files, dropping dummy row
# files1 <- list.files(areaDir, pattern="*_test5_RRCA_POLYGRID_1mile.csv")
# area1 = loadStats(files1)[-1,] %>%
#   dplyr::select(layer, year, pr_ann, irrigated_m2) %>%
#   rename(NODE = layer) %>%
#   mutate(unique2 = paste0(year, '-', NODE))

# load 4 mile files, dropping dummy row
files4 <- list.files(areaDir,
                     pattern="*_test5_RRCA_POLYGRID_4mile_weighted.csv")
area4 = loadStats(files4)[-1,] %>%
  dplyr::select(layer, year, pr_ann, irrigated_m2) %>%
  rename(NODE = layer) %>%
  mutate(unique2 = paste0(year, '-', NODE)) 
```

# Calculate depth
Match up datasets and calculate depth. I wrote out my generated raster/polygon and confirmed the 'layer' variable (cell number) matched the shapefile for pumping data 'node' column

## 4 mile RRCA grid

### Match up datasets
The 4 mile grid aggregates 4 RRCA cells, which are nonsequential. Get the 4 NODE values per polygon.

```{r getNodes4, cache=TRUE}
# load the 4 mile polygrid
outPolyDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/kml/FINAL_BOUNDARIES'
poly4 <- readOGR(outPolyDir, 'RRCA_polygrid_4mile', verbose=FALSE)

# remake the 1 mile raster with grid cell numbers
CRS <- '+proj=utm +zone=14 +ellps=clrk66 +datum=NAD27 +units=us-ft +no_defs'
template <- raster(crs=CRS, nrows = 165, ncols=326, resolution = 5280,
                 ymn = 14092806, ymx = 14964006, 
                 xmn = 266023, xmx = 1987023)
template[] <- 1:53790 

# get the raster values per polygon
poly4$ID <- 1:nrow(poly4)
nodeLookup <- raster::extract(template, poly4, df=TRUE)

# add a column back in for the node/layer value of the polygrid 
nodeLookup2 <- nodeLookup %>%
  rename(NODE = layer) %>%
  full_join(poly4@data)

# so the following NODES from the irrigation water data are associated with 
# aggregated cell "2725"
nodeLookup2 %>% filter(layer == 2725)
```

Awesome. Now, join the irrigation water data to this lookup table and aggregated to layer (poly4 )cells.

Note due to using the polygon buffer to "clip" the 4 mile grid raster, there are some poly4 cells 100% outside of the RRCA model grid (and hence have NA values; remove them)

```{r aggregateWater}
# aggregate irrigation water by 4 mile polygons for each year
irrBy4 <- nodeLookup2 %>%
  full_join(irrByCell) %>%                       # join irrigation water dataset
  filter(!is.na(totalIrr_m3)) %>%                # remove NAs
  group_by(year, layer) %>%                      
  summarise(totalIrr_m3 = sum(totalIrr_m3)) %>%  # get total for 16 model cells
  rename(NODE = layer)                           # rename to match area df
#looky
irrBy4 %>% filter(NODE==2725)
```

### calculate depth
join irrigation volume df with area df and calculate depth already

```{r depth}
# combine datasets by year + node/layer and calculate depth
depth4 <- irrBy4 %>%
  left_join(area4) %>%
  mutate(depth_cm = totalIrr_m3/irrigated_m2 * 100) %>%
  filter(is.finite(depth_cm))

# check out the depth data range
summary(depth4$depth_cm)

# good quantiles for cleaning?
quantile(depth4$depth_cm, c(.9,.99))

# filter for the 90% quantile and lower
depth4.90 <- depth4 %>% filter(depth_cm < 107)
hist(depth4.90$depth_cm)
boxplot(depth4.90$depth_cm)
```

### plots and analysis

#### Depth Trend
Calculate and map the trend in depth over time, accounting for precip. Model form: depth_cm ~ year + pr_ann

Plot: the coefficient for year

```{r trend4}
# function to get significant year coeffs by cell
getSpatialRegression <- function(dataset, polygrid, p = 0.05){
  # run regressions
  lms <- dataset %>% 
      mutate(pr_ann_cm = pr_ann/10) %>%   # convert precip to cm
      group_by(NODE) %>% 
      do(model = lm(depth_cm ~ year + pr_ann_cm, data = .))
 
  # tidy up with broom
  coeffs <- lms %>% tidy(model)
  coeffs <- coeffs[coeffs$term == 'year',]
 
   # merge coefficients with grid polygons
  poly <- sp::merge(polygrid, coeffs[,c('NODE','estimate','p.value')])
  # just sig trends
  poly[is.na(poly$p.value),'p.value'] <- 1 # dummy p value no data county
  poly$sigCoeffs <- poly$estimate
  poly[poly$p.value >= p, 'sigCoeffs'] <- NA
  return(poly)
} 

# add a node layer
poly4$NODE <- poly4$layer
# run for 4 miles
trends4 <- getSpatialRegression(depth4.90, poly4)
# plot
spplot(trends4, 'sigCoeffs', main = 'Depth trends...by year?!')
```

And make a trend map

```{r trendMap}
#nice plotting -------------------------
# polygon boundaries
gisDir <- "C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/kml/FINAL_BOUNDARIES"
rrbpoly <- readOGR(gisDir, 'RRB_aea', verbose=F)
rrca <- readOGR(gisDir, 'RRCA_modified_aea', verbose=F)
# get union bounds
unioned <- readOGR(gisDir, 'RRCA_RRB_union_aea', verbose=F)

states <- readOGR('C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/kml/shapefile', 'States', verbose=F)
states <- states[states$STATE_ABBR %in% c('CO','NE','KS'),]
states <- spTransform(states, CRS(proj4string(rrca)))

# palette parameters
# blue and red color ramps pulled from color brewer (no white)
redRamp2 <- colorRampPalette(c('#fddbc7','#f4a582','#d6604d','#b2182b','#67001f'))
blueRamp2 <- colorRampPalette(rev(c('#d1e5f0','#92c5de','#4393c3','#2166ac','#053061')))

colorbreaks <- 200
biPal <- c(blueRamp2(colorbreaks/2),redRamp2(colorbreaks/2))

# scale range
palMin <- -12
palMax <- 12

# reproject trends4
trends4.aea <- spTransform(trends4, crs(proj4string(rrca)))

# gray background
spplot(trends4.aea, 'sigCoeffs', col.regions=biPal, col=NA,
       at=seq(palMin,palMax,(-palMin+palMax)/colorbreaks),
              main = 'Trends in Depth (cm/year)',
       colorkey=list(space='bottom',
                     labels=list(
                       labels = c('-12','-8','-4','0','4','8','12'),
                       at = c(-12,-8,-4,0,4,8,12)))) +
  latticeExtra::layer(sp.polygons(rrbpoly, lty=2, lwd=.8)) + 
  latticeExtra::layer(sp.polygons(rrca,lwd=.8)) +
  latticeExtra::layer(sp.polygons(states, lwd=1, col='gray30')) + 
  latticeExtra::layer(sp.polygons(rrca, fill = 'gray50', col=NA), under=T)

```

#### Mean Depth
Calculate and map the mean depth in space

```{r mean4, cache=TRUE}
# get mean depth per year for Xiao
depth4.mean <- depth4.90  %>%
  group_by(NODE) %>%
  summarize(meanDepth_cm = mean(depth_cm))

summary(depth4.mean$meanDepth_cm)

# join the data
poly4.meanDepth <- merge(poly4, depth4.mean, by.x = 'layer', by.y = 'NODE')

# rasterize for better plotting
meanDepth_4mile <- rasterize(poly4.meanDepth, template, 
                             field = 'meanDepth_cm')
spplot(meanDepth_4mile, main = 'Mean Irrigation Depth, 1999-2015')

# try an sf plot
meanDepth4.sf <- st_as_sf(poly4.meanDepth)
plot(meanDepth4.sf['meanDepth_cm'], border = NA, main = 'Mean Depth 1999-2015, 4m')

# area trends stle plot -------------------------------------

# palette parameters
# blue and red color ramps pulled from color brewer (no white)
redRamp2 <- colorRampPalette(c('#fddbc7','#f4a582','#d6604d','#b2182b','#67001f'))
blueRamp2 <- colorRampPalette(rev(c('#d1e5f0','#92c5de','#4393c3','#2166ac','#053061')))

colorbreaks <- 200
biPal <- c(blueRamp2(colorbreaks/2),redRamp2(colorbreaks/2))

# scale range
palMin <- -12
palMax <- 12

# reproject trends4
poly4.meanDepth.aea <- spTransform(poly4.meanDepth, crs(proj4string(rrbpoly)))

# gray background
spplot(poly4.meanDepth.aea, 'meanDepth_cm', col.regions=biPal, col=NA,
       #at=seq(palMin,palMax,(-palMin+palMax)/colorbreaks),
              #main = 'Absolute area, 4 km grid, sig only',
       colorkey=list(space='bottom') + #,
                     #labels=list(
                    #   labels = c('-12','-8','-4','0','4','8','12'),
                    #   at = c(-12,-8,-4,0,4,8,12)))) +
  latticeExtra::layer(sp.polygons(rrbpoly, lty=2, lwd=.8)) + 
  latticeExtra::layer(sp.polygons(rrca,lwd=.8)) +
  latticeExtra::layer(sp.polygons(states, lwd=1, col='gray30')) + 
  latticeExtra::layer(sp.polygons(unioned, fill = 'gray50', col=NA), under=T))


# # write out 
# outRas <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/pumping/depthRasters'
# writeRaster(meanDepth_4mile, 
#             paste0(outRas,'/RRCA_4mile_1999-2015_meanDepth_cm_90p.tif'))
```




## 1 mile grid
To check with Xiao's

```{r 1mile, cache=TRUE}
# load 1 mile files, dropping dummy row
files1 <- list.files(areaDir, pattern="*_test5_RRCA_POLYGRID_1mile.csv")
area1 = loadStats(files1)[-1,] %>%
  dplyr::select(layer, year, pr_ann, irrigated_m2) %>%
  rename(NODE = layer) 

# combine datasets by year + node/layer and calculate depth
depth1 <- irrByCell %>%
  left_join(area1) %>%
  mutate(depth_cm = totalIrr_m3/irrigated_m2 * 100) %>%
  filter(is.finite(depth_cm))

# check out the depth data range - raw
summary(depth1$depth_cm)

# good quantiles for cleaning?
quantile(depth1$depth_cm, c(.9,.99))

# filter for irrigation < 80 cm to match Xiao
depth1.Xiao <- depth1 %>% filter(depth_cm < 80)
summary(depth1.Xiao$depth_cm)
hist(depth1.Xiao$depth_cm)
boxplot(depth1.Xiao$depth_cm)

# get mean depth per year for Xiao
depth1.mean <- depth1.Xiao %>%
  group_by(NODE) %>%
  summarize(meanDepth_cm = mean(depth_cm))

summary(depth1.mean$meanDepth_cm)

# load the 1 mile polygrid
outPolyDir <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/GIS/kml/FINAL_BOUNDARIES'
poly1 <- readOGR(outPolyDir, 'RRCA_polygrid_1mile', verbose=FALSE)

# join the data
poly1.meanDepth <- merge(poly1, depth1.mean, by.x = 'layer', by.y = 'NODE')

# rasterize for better plotting
meanDepth_1mile <- rasterize(poly1.meanDepth, template, field = 'meanDepth_cm')
spplot(meanDepth_1mile)

# write out for Xiao
outRas <- 'C:/Users/deinesji/Dropbox/1PhdJill/hpa/irrigation/data/pumping/depthRasters'
writeRaster(meanDepth_1mile, paste0(outRas,
                                    '/RRCA_1mile_1999-2015_meanDepth_cm.tif'))
```
